% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\encoding{UTF-8}
\name{age}
\alias{age}
\title{Métrique}
\usage{
age(naissance, reference)
}
\arguments{
\item{naissance}{Date de naissance de l'individu au format AAAA-MM-JJ.}

\item{reference}{Date de référence à laquelle on veut calculer l'âge de l'individu.}
}
\value{
\code{vector}
}
\description{
Calcul de l'âge d'un individu à une date de référence.
}
\details{
Il existe plusieurs manière de calculer l'âge dont diviser par 365 ou 365.25. Ces divisions ne gèrent pas les années bissextiles alors que l'utilisation de la fonction \code{age()} le fait.
}
\examples{
library(inesss)

# Vecteur
vec_age <- age(ech_population$NAISS, Sys.Date())

# dplyr
library(dplyr)
df <- ech_population \%>\% mutate(AGE = age(NAISS, Sys.Date()))
\dontrun{
# data.table
library(data.table)
dt <- copy(ech_population)
dt[, AGE := age(NAISS, Sys.Date())]
}
}
