% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{as_date_excel_chr}
\alias{as_date_excel_chr}
\title{Utils}
\usage{
as_date_excel_chr(x)
}
\arguments{
\item{x}{Vecteur contenant les dates Excel au format entier.}
}
\description{
Convertir une date Excel au format CHR en Date R. Soit une valeur 'AAAA-MM-JJ'
ou un nombre entier, mais au format CHR (spécifié par l'utilisateur).
}
\keyword{internal}
