% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{Charlson_Dx_UManitoba16}
\alias{Charlson_Dx_UManitoba16}
\title{Data - Codes diagnostics}
\format{
\code{list(Dx = list(CIM9, CIM10))}
}
\source{
\href{http://mchp-appserv.cpe.umanitoba.ca/reference/Candata_web_final.pdf#Page=92}{CANCER DATA LINKAGE IN MANITOBA: EXPANDING THE INFRASTRUCTURE FOR RESEARCH} page 72 du document.
}
\usage{
data('Charlson_Dx_UManitoba16')
}
\description{
Codes SQL regex (se terminent par un '\%') à utiliser lors de l'extraction des codes de diagnostics pour l'étude de la comorbidité.
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.\cr\cr
\code{aids} : HIV/AIDS\cr
\code{canc} : Cancer\cr
\code{chf} : Congestive Heart Failure\cr
\code{cpd} : Chronic Pulmonary Disease\cr
\code{ctdrd} : Connective Tissue Disease - Rheumatic Disease\cr
\code{cvd} : Cerebrovascular Disease\cr
\code{dementia} : Dementia\cr
\code{diab} : Diabetes with Chronic Complications\cr
\code{diabwc} : Diabetes without Chronic Complications\cr
\code{ld1} : Mild Liver Disease\cr
\code{ld2} : Moderate or Severe Liver Disease\cr
\code{mc} : Metastatic Carcinoma\cr
\code{mi} : Myocardial Infarction\cr
\code{ph} : Paraplegia and Hemiplegia\cr
\code{pud} : Peptic Ulcer Disease\cr
\code{pvd} : Peripheral Vascular Disease\cr
\code{rd} : Renal Disease
}
\keyword{datasets}
