% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_vec.R
\name{chunk_vec}
\alias{chunk_vec}
\title{Tronquer un vecteur en plusieurs parties}
\usage{
chunk_vec(x, n_chunks = NULL, n_vals = NULL)
}
\arguments{
\item{x}{Vecteur à tronquer en plusieurs parties.}

\item{n_chunks}{Diviser le vecteur en \code{n_chunks} parties.}

\item{n_vals}{Chaque partie aura au maximum \code{n_vals} valeurs.}
}
\value{
\code{list} ayant \code{n_chunks} éléments (ou \code{as.integer(length(x) / n_vals + 1L)}).
}
\description{
Divise le vecteur \code{x} en \code{n_chunks} parties ou le divise pour avoir au maximum \code{n_vals} valeurs dans chaque partie.
}
\details{
Utiliser l'argument \code{n_chunks} ou \code{n_vals}, pas les deux.
}
\examples{
chunk_vec(x = 1:10, n_chunks = 3)
chunk_vec(x = 1:10, n_vals = 3)
}
