% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{CIM9}
\alias{CIM9}
\title{Table ou Liste}
\format{
Tableau de 2 variables et 7184 observations :
\describe{
\item{CODE}{Code de diagnostic CIM-9. \code{character}.}
\item{DIAGNOSTIC}{Description du code de diagnostic. \code{character}.}
}
}
\source{
\href{https://www.ramq.gouv.qc.ca/fr/professionnels/medecins-omnipraticiens/facturation/repertoire-diagnostics/Pages/repertoire-diagnostics.aspx}{Répertoire des diagnostics}.
}
\usage{
data('CIM9')
}
\description{
Version légèrement modifiée par la RAMQ pour la facturation.
}
\keyword{datasets}
