% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{CIM_correspond}
\alias{CIM_correspond}
\title{Table ou Liste}
\format{
Tableau de 4 variables et 25866 observations :
\describe{
\item{CIM9}{Code de diagnostic CIM-9. \code{character}.}
\item{CIM9_DESC}{Description du code de diagnostic. \code{character}.}
\item{CIM10}{Code de diagnostic CIM-10. \code{character}.}
\item{CIM10_DESC}{Description du code de diagnostic. \code{character}.}
}
}
\source{
\href{https://www.ramq.gouv.qc.ca/fr/professionnels/medecins-omnipraticiens/facturation/repertoire-diagnostics/Pages/repertoire-diagnostics.aspx}{Répertoire des diagnostics}.
}
\usage{
data('CIM_correspond')
}
\description{
Tableau de correspondance entre la CIM-9 et la CIM-10
}
\keyword{datasets}
