% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiner_periodes.R
\encoding{UTF-8}
\name{combiner_periodes}
\alias{combiner_periodes}
\title{Combinaisons périodes d'études}
\usage{
combiner_periodes(dt, id, debut, fin, par_cols = NULL, njours = 1)
}
\arguments{
\item{dt}{Data d'analyse.}

\item{id}{Nom de la colonne indiquant l'identifiant de l'individu.}

\item{debut}{Nom de la colonne indiquant le début de la période.}

\item{fin}{Nom de la colonne indiquant la fin de la période.}

\item{par_cols}{Nom des autres colonnes qui doivent être incluses dans l'analyse. Par exemple des codes de médicaments. Par défaut \code{NULL}.}

\item{njours}{Nombre de jours max entre le début et la fin précédente pour effectuer une combinaison.}
}
\description{
Combine, fusionne des périodes qui se chevauchent dans le temps.
}
