% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\encoding{UTF-8}
\name{comorbidity}
\alias{comorbidity}
\title{Comorbidity}
\usage{
comorbidity(
  dt,
  ID,
  DIAGN,
  DATE_DX,
  SOURCE,
  n1 = 30,
  n2 = 730,
  method = c("Charlson", "Elixhauser"),
  scores = "CIM10",
  confirm_sourc = list(`MED-ECHO` = 1, BDCU = 2, SMOD = 2)
)
}
\arguments{
\item{dt}{Dataset ayant au moins les quatre (4) colonnes \code{ID}, \code{DIAGN}, \code{DATE_DX} et \code{SOURCE}.}

\item{ID}{Nom de la colonne indiquant le numéro de l'usager, de l'individu.}

\item{DIAGN}{Nom de la colonne indiquant le code d'un diagnostic. Voir \code{names(inesss::Comorbidity_diagn_codes)}.}

\item{DATE_DX}{Nom de la colonne indiquant la date du diagnostic.}

\item{SOURCE}{Nom de la colonne indiquant la provenance du diagnostic.}

\item{n1, n2}{Nombre de jours dans le but de construire l'intervalle \verb{[n1,n2]}. Pour qu'un code de diagnostique soit confirmé, il faut que \emph{DIAGN{i}} soit suivi de \emph{DIAGN{j}} (où i < j) et que le nombre de jours entre les deux soit dans l'intervalle \verb{[n1,n2]}.}

\item{method}{Méthode de calcul des indicateurs. \code{'Charlson'} ou \code{'Elixhauser'}. Inscrire les deux crée la colonne \code{Combined}.}

\item{scores}{\code{'CIM9'} ou \code{'CIM10'}. Nom de la colonne du dataset \code{Comorbidity_weights} à utiliser pour le calcul des indicateurs.}

\item{confirm_sourc}{\code{list} indiquant la \emph{confiance} des \code{SOURCE}. Si une \code{SOURCE} doit être confirmé par un autre code dans l'intervalle \verb{[n1,n2]}, inscrire \code{2}, sinon \code{1}. Inscrire les sources sous le format : \code{confirm_sourc = list(source1 = 1, source2 = 2, source3 = 2, ...)}. \code{confirm_sourc} doit contenir toutes les valeurs uniques de la colonne \code{SOURCE}.}
}
\value{
\code{data.table} avec les colonnes \code{ID}, \code{Charlson} (selon \code{method}), \code{Elixhauser} (selon \code{method}), \code{Combined} (selon \code{method}), et tous les codes de diagnostiques indiquant leur poids.
}
\description{
Calcul des indicateurs de \emph{Charlson}, \emph{Elixhauser} et la combinaison des deux.
}
\details{
\strong{\code{confirm_sourc} :} Dans l'exemple \code{confirm_sourc = list(source1=1, source2=2, source3=2, ...)}, la \code{source3} pourrait confirmer la \code{source2} et vice-versa.
}
