% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{Comorbidity_weights}
\alias{Comorbidity_weights}
\title{Data - Poids des codes de diagnostics}
\format{
Tableau de 4 variables et 32 observations.
\describe{
\item{DIAGN}{Identification du diagnostic. \code{character}.}
\item{DIAGN_CODE}{Code de diagnostic utilisé utilisé dans la programmation. \code{character}.}
\item{CIM9}{Poids utilisé pour la \emph{9e} révision du CIM. \code{integer}.}
\item{CIM10}{Poids utilisé pour la \emph{10e} révision du CIM. \code{integer}.}
}
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/29578951/}{Validation of the Combined Comorbidity Index of Charlson and Elixhauser to Predict 30-Day Mortality Across ICD-9 and ICD-10}. \href{https://github.com/INESSS-QC/inesss1/blob/master/inst/articles/CCI_INSPQ.pdf}{Voir PDF.}
}
\usage{
data('Comorbidity_weights')
}
\description{
Data - Poids des codes de diagnostics
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
