% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{ComorbidityWeights}
\alias{ComorbidityWeights}
\title{Data - Poids des codes de diagnostics}
\format{
\code{list} contenant trois (3) tables indiquant la description, le code et le poids :
\describe{
\item{CCI_INSPQ_2018_CIM9}{}
\item{CCI_INSPQ_2018_CIM10}{}
\item{UManitoba_2016}{}
}
}
\source{
Voir la source des datas \code{\link{Combine_Dx_CCI_INSPQ18}}, \code{\link{Charlson_Dx_CCI_INSPQ18}}, \code{\link{Elixhauser_Dx_CCI_INSPQ18}} et \code{\link{Charlson_Dx_UManitoba16}}
}
\usage{
data('ComorbidityWeights')
}
\description{
Data - Poids des codes de diagnostics
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
