% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_Dx_by_nDx.R
\encoding{UTF-8}
\name{confirm_Dx_par_nDx}
\alias{confirm_Dx_par_nDx}
\title{Confirmation Diagnostic}
\usage{
confirm_Dx_par_nDx(
  dt,
  ID,
  DIAGN,
  DATE_DX,
  nDx = 1,
  n1 = 30,
  n2 = 730,
  keep_all = FALSE
)
}
\arguments{
\item{dt}{Table d'analyse.}

\item{ID}{Nom de la colonne indiquant le numéro de l'usager, de l'individu.}

\item{DIAGN}{Nom de la colonne indiquant le numéro de l'usager, de l'individu.}

\item{DATE_DX}{Nom de la colonne indiquant la date du diagnostic.}

\item{nDx}{Nombre de diagnostics qui doivent confirmer une date précédente.}

\item{n1, n2}{Nombre de jours requis pour confirmer le diagnostic}

\item{keep_all}{Afficher toutes les dates qui sont suivies de \code{nDx} date(s) dans un intervalle de [\code{n1}, \code{n2}] jours (\code{TRUE}), ou conserver seulement la première date (\code{FALSE}).}
}
\value{
\code{data.table}
}
\description{
Indique la liste des diagnostics qui sont confirmés par \code{nDx} dans un intervalle de [\code{n1}, \code{n2}] jours.
}
