% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_nDx.R
\encoding{UTF-8}
\name{confirm_nDx}
\alias{confirm_nDx}
\alias{confirm_2Dx}
\alias{confirm_3Dx}
\title{Confirmation Diagnostics}
\usage{
confirm_2Dx(
  dt,
  ID,
  DATE,
  DIAGN = NULL,
  study_start = NULL,
  study_end = NULL,
  n1 = 30,
  n2 = 730,
  reverse = FALSE
)

confirm_3Dx(
  dt,
  ID,
  DATE,
  DIAGN = NULL,
  study_start = NULL,
  study_end = NULL,
  n1 = 30,
  n2 = 730,
  reverse = FALSE
)
}
\arguments{
\item{dt}{Table contenant les dates de diagnostics des individus.}

\item{ID}{Nom de la colonne contenant le numéro d'identification unique des individus.}

\item{DATE}{Nom de la colonne contenant la date du diagnostic.}

\item{DIAGN}{Facultatif. Nom de la colonne indiquant les codes de diagnostics.}

\item{study_start}{Date de début de la période d'étude contenant les dates de repérage. Si \code{NULL}, aura pour valeur la première date de \code{dt}, la plus ancienne.}

\item{study_end}{Date de fin de la période d'étude contenant les dates de repérage. Si \code{NULL}, aura pour valeur la dernière date de \code{dt}, la plus récente.}

\item{n1, n2}{Nombre de jours permettant de construire l'intervalle [n1; n2] où un code de diagnostic peut en confirmer un autre.}

\item{reverse}{\code{TRUE} ou \code{FALSE}. Si on doit faire la vérification en prenant la date la plus récente et en reculant dans le temps.}
}
\description{
Confirmation d'un diagnostic par d'autres diagnostics lorsque ceux-ci se retrouvent dans un intervalle précis.
}
