% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{conn}
\alias{conn}
\title{Utils}
\usage{
conn(dsn, uid, pwd)
}
\arguments{
\item{dsn}{\strong{D}ata \strong{S}ource \strong{N}ame.}

\item{uid}{User Identifier.}

\item{pwd}{Password to use.}
}
\description{
Connexion entre R et SQL Teradata
}
