% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ymd.R
\name{date_ymd}
\alias{date_ymd}
\title{Date}
\usage{
date_ymd(yyyy, mm, dd)
}
\arguments{
\item{yyyy}{Nombre entier indiquant l'année.}

\item{mm}{Nombre entier compris entre 1 et 12, où 1 indique janvier et 12 décembre.}

\item{dd}{Nombre compris entre 1 et 31 selon les mois. Pour remplacer le dernier jour du mois (28, 29, 30, 31), il est possible d'inscrire \code{dd = 'last'}.}
}
\value{
lubridate::as_date
}
\description{
Retourne une date au format \code{AAAA-MM-JJ}. Utile dans des \verb{for loop}, car \code{dd} peut prendre la valeur \code{'last'} (au lieu d'un nombre), donc pas besoin de savoir si le dernier jour du mois est le 28 ou le 29 en février, ou un 30 ou un 31 pour les autres mois.
}
\examples{
date_ymd(2020, 1, 15)
date_ymd(2020, 10, 31)
date_ymd(2020, 6, 'last')
for (yr in 1996:2004) {
  print(date_ymd(yyyy = yr, mm = 2, dd = 'last'))
}
}
