% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{DBC_TablesV}
\alias{DBC_TablesV}
\title{Domaine de valeur}
\format{
\code{data.table} de 3 variables et 32708 observations
\describe{
\item{NOM_BD}{Nom de la base de données. Par exemple \emph{PROD}, \emph{RES_SSS}, \emph{DONNE_INESSS}, etc. \code{character}.}
\item{NOM_TABLE}{Nom de la table. \code{character}.}
\item{TYPE_TABLE}{Type de la table. Voir \href{https://docs.teradata.com/r/EcH18cz_mkFcqxA7GeutqQ/kD2HkTfj10WlyDqsbQrc4w}{TableKind Column} pour plus d'infos. \code{character}.}
}
}
\usage{
data('DBC_TablesV')
}
\description{
Domaine de valeur
}
\keyword{internal}
