% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DER_SQL_generateur.R
\encoding{UTF-8}
\name{DER_SQL_generateur}
\alias{DER_SQL_generateur}
\title{SQL générateur}
\usage{
DER_SQL_generateur()
}
\value{
\code{shiny}
}
\description{
Générateur de requêtes SQL pour la DER.
}
\details{
L'utilisateur sélectionne une méthode et la requête SQL est générée. Il peut ensuite modifier la requête directement ou simplement inscrire les valeurs des variables dans l'espace dédié aux arguments.
}
