% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins-rstudio.R
\encoding{UTF-8}
\name{DER_update_package}
\alias{DER_update_package}
\title{Update inesss package}
\usage{
DER_update_package()
}
\value{
\link{install_local_inesss}
}
\description{
Update inesss package
}
\keyword{internal}
