% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domaine_valeurs.R
\encoding{UTF-8}
\name{domaine_valeurs}
\alias{domaine_valeurs}
\title{Domaine de valeurs}
\usage{
domaine_valeurs()
}
\description{
Domaine de valeurs de différentes bases de données disponibles dans le domaine de la santé, principalement à la RAMQ.
}
