% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins-rstudio.R
\encoding{UTF-8}
\name{domaine_valeurs_addin}
\alias{domaine_valeurs_addin}
\title{Démarrer domaine de valeur}
\usage{
domaine_valeurs_addin()
}
\value{
\link{domaine_valeurs}
}
\description{
Démarrer domaine de valeur
}
\keyword{internal}
