% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ech_population.R
\docType{data}
\name{ech_population}
\alias{ech_population}
\title{DATA}
\format{
Un \code{data.table} avec 10000 lignes et 3 colonnes.
}
\usage{
data(ech_population)
}
\description{
\code{data.table} contenant les informations d'une population.
}
\details{
\itemize{
\item ID. Numéro d'identification de l'individu. (1--10000)
\item SEXE. Sexe de l'individu. (F (49.5\%), H (49.5\%), X (1.0\%))
\item NAISS. Date de naissance. ("1940-01-01" -- "1980-12-31")
}
}
\keyword{datasets}
