% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{file_directory}
\alias{file_directory}
\title{File status}
\usage{
file_directory()
}
\value{
CHR. Emplacement du dossier qui contient le script R.
}
\description{
Emplacement sur le disque dur où le script est sauvegardé.
}
\details{
Si le script n'est pas sauvegardé, retourne \code{NULL}.
}
\examples{
file_directory()
}
