% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\encoding{UTF-8}
\name{indent}
\alias{indent}
\title{Utils}
\usage{
indent(niv = 1)
}
\arguments{
\item{niv}{Niveau d'indentation.}
}
\value{
Quatre (4) espaces répétés \code{niv} fois.
}
\description{
Indentation du code
}
\keyword{internal}
