% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inesss_excel_macros.R
\encoding{UTF-8}
\name{inesss_excel_macros}
\alias{inesss_excel_macros}
\title{Macros VBA}
\usage{
inesss_excel_macros(filename = "PERSONAL.XLSB")
}
\arguments{
\item{filename}{Nom du fichier contenant les macros VBA. Si le fichier existe déjà, on demande si on veut le remplacer. Par défaut \code{'PERSONAL.XLSB'}.}
}
\description{
Télécharge le fichier Excel contenant les macros VBA et l'enregistre au répertoire suivant :\cr
\code{C:\\Users\\NomUtilisateur\\AppData\\Roaming\\Microsoft\\Excel\\XLSTART}.\cr
L'enregistrer à cet endroit permet de rendre disponible les macros VBA à partir de n'importe quelle classeur Excel.
}
\examples{
\dontrun{
inesss_excel_macros()  # nom = 'PERSONAL.XLSB'
inesss_excel_macros(filename = 'LeNomQueJeVeux.xlsb')
}
}
