% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_local_inesss.R
\name{install_local_inesss}
\alias{install_local_inesss}
\title{Install package inesss}
\usage{
install_local_inesss(direction)
}
\arguments{
\item{direction}{Nom de la direction.}
}
\description{
Installation ou mise à jour de la librairie selon la direction indiquée.
}
