% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_RDCOMClient.R
\encoding{UTF-8}
\name{install_RDCOMClient}
\alias{install_RDCOMClient}
\title{Package - Installation}
\usage{
install_RDCOMClient(msg = TRUE)
}
\description{
Installation de la librairie \href{https://github.com/omegahat/RDCOMClient}{RDCOMCLIENT}. Si elle est déjà installée, le programme demande si on veut la mettre à jour.
}
