% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{list_as_dt_no_recycled}
\alias{list_as_dt_no_recycled}
\title{Utils}
\usage{
list_as_dt_no_recycled(List)
}
\arguments{
\item{List}{Liste à convertir en data.table}
}
\description{
Transforme une \code{list} en \code{data.table} sans répéter les valeurs, ajoute des \code{NA}s.
}
\keyword{internal}
