% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\encoding{UTF-8}
\name{nl}
\alias{nl}
\title{Utils}
\usage{
nl(x = 1)
}
\arguments{
\item{x}{Nombre de fois que l'on répète le retour de ligne.}
}
\value{
'\\n'
}
\description{
Utils
}
\keyword{internal}
