% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_type_rx.R
\encoding{UTF-8}
\name{nom_type_rx}
\alias{nom_type_rx}
\title{Utils}
\usage{
nom_type_rx(type_rx)
}
\arguments{
\item{type_rx}{Type de code, "DENOM", "DIN", ...}
}
\value{
\code{"DENOM"} : \code{"NOM_DENOM"}.\cr
\code{"DIN"} = \verb{"NOM_MARQ_COMRC}.
}
\description{
Indique le nom de la colonne indiquant le nom du type de Rx dans un data
}
\keyword{internal}
