% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{Obstetrics_Dx}
\alias{Obstetrics_Dx}
\title{Data - Codes diagnostics gestationnels}
\format{
\code{list(Dx = list(CIM9, CIM10))}
}
\usage{
data('Obstetrics_Dx')
}
\description{
Codes SQL regex (se terminent par un '\%') à utiliser lors de l'extraction des codes de diagnostics gestationnels pour l'étude de la comorbidité.
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
