% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook_mail.R
\encoding{UTF-8}
\name{outlook_mail}
\alias{outlook_mail}
\title{Courriel}
\usage{
outlook_mail(
  to = NULL,
  cc = NULL,
  subject = NULL,
  body = NULL,
  attachments = NULL
)
}
\arguments{
\item{to}{Destinataire(s). Section \code{From} / \code{À}.}

\item{cc}{Destinaire(s). Section \code{Cc}.}

\item{subject}{Objet du courriel.}

\item{body}{Message du courriel. Utiliser \verb{\\n} pour un retour de ligne.}

\item{attachments}{Répertoire du ou des fichiers à mettre en pièce jointe.}
}
\description{
Envoyer un courriel à partir de Outlook.\cr
\strong{ATTENTION} Vérifier l'adresse utilisée s'il y a plusieurs comptes.
}
\examples{
outlook_mail(to = "Mon.Directeur@inesss.qc.ca",
             cc = "Mon.Collegue@inesss.qc.ca",
             subject = "Projet1 - Finalisation",
             body = "Bonjour,\n Il faudrait se rencontrer pour en discuter.\n
                     Merci\n\n Mon Nom")
}
