% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{Pop_QC}
\alias{Pop_QC}
\title{Data - Estimations et projections de population comparables (1996-2041)}
\format{
Tableau de 8 variables et 2 595 320 observations :
\describe{
\item{GEO}{Niveau géographique : Québec, RUISSS, RSS, RTS, RLS, CLSC.}
\item{CODE}{Code du territoire.}
\item{AN}{Année.}
\item{TYPE}{Type de données : Estimations ou Projections.}
\item{STATUT}{Donnée révisée ou provisoire. \code{NA} indique que la donnée n'a pas été changée depuis la dernière publication.}
\item{SEXE}{}
\item{AGE}{}
\item{POP}{Population}
}
}
\source{
\href{https://www.msss.gouv.qc.ca/professionnels/informations-geographiques-et-de-population/donnees-populationnelles/}{MSSS Données de population}.\cr
\href{https://github.com/INESSS-QC/inesss1/blob/master/inst/extdata/EstimationProjectionComparable_1996_2041_20200424.xlsx}{Fichier Excel utilisé}.
}
\usage{
data('Pop_QC')
}
\description{
Tableau de la population québécoise par niveau géographique.\cr\cr
Ce fichier présente une série continue de données populationnelles comparables composée de la série des estimations (1996-2019) et de la série des projections (2020-2041) de population. Ces données tiennent compte de l'évolution de la population selon les plus récentes données observées de naissances, décès et mouvements migratoires.\cr\cr
Il est à noter que ces données de population sont présentées sur la base du découpage territorial du réseau de la santé et des services sociaux, soit pour les territoires suivants : le Québec, les réseaux universitaires intégrés de santé et de services sociaux (RUISSS), les régions sociosanitaires (RSS), les réseaux territoriaux de services (RTS), les réseaux locaux de services (RLS) et les centres locaux de services communautaires (CLSC).
}
\details{
\strong{Attention} \code{AGE = 90} équivaut à \emph{90 ans et plus}.\cr\cr
La classe des colonnes est \code{character} lorsque c'est du texte ou \code{integer} lorsque c'est un nombre.\cr\cr
\strong{Mise en ligne} : 25 février 2016.\cr
\strong{Dernière modification} : 24 avril 2020.\cr
\strong{Publication no} : EstimProjComp-ISQ.\cr\cr
La fiche d'information et technique de cette base de données est disponible avec le fichier Excel (voir \emph{Source}).\cr\cr
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
