% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.statistiques_generales.method.R
\encoding{UTF-8}
\name{query.statistiques_generales.method}
\alias{query.statistiques_generales.method}
\title{SQL requête}
\usage{
query.statistiques_generales.method(
  debut,
  fin,
  type_Rx,
  codes = NULL,
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  group_by = NULL,
  cohort = FALSE
)
}
\arguments{
\item{debut}{Date de début au format AAAA-MM-JJ. \code{string}.}

\item{fin}{Date de fin au format AAAA-MM-JJ. \code{string}.}

\item{type_Rx}{Type de médicament : {"DENOM", "DIN"}. \code{string}.}

\item{codes}{Codes à analyser. \code{string} ou \code{numeric}.}

\item{code_serv}{Liste des codes de services à inclure ou exclure selon \code{code_serv_filtre}. \code{string}.}

\item{code_serv_filtre}{"Exclusion" ou "Inclusion". \code{string}.}

\item{group_by}{Variables dont on doit aggréger les résultats. \code{string}.}

\item{cohort}{Vrai ou Faux. Si la requête est dépendante d'une cohorte. \code{logical}.}
}
\value{
\code{string}
}
\description{
Statistiques générales
}
\keyword{internal}
