% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM.R
\encoding{UTF-8}
\name{query_I_SMOD_SERV_MD_CM}
\alias{query_I_SMOD_SERV_MD_CM}
\title{Code SQL}
\usage{
query_I_SMOD_SERV_MD_CM(debut, fin, diagn)
}
\arguments{
\item{debut}{Date de début de la période d'étude.}

\item{fin}{Date de fin de la période d'étude.}

\item{diagn}{\code{vector} indiquant les codes de diagnostics \emph{CIM9} et/ou \emph{CIM10} à extraire.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
Générateur de code SQL pour l'extraction de diagnostics de la vue \code{I_SMOD_SERV_MD_CM}.
}
\keyword{internal}
