% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch.old_user_where_codes}
\alias{query_naif_switch.old_user_where_codes}
\title{query_naif_switch1}
\usage{
query_naif_switch.old_user_where_codes(type_Rx_retro, rx_retrospect_a_exclure)
}
\description{
Générateur de code SQL pour la section du WITH \emph{OLD_USER}, \emph{where}, \emph{codes}.
}
\keyword{internal}
