% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch1}
\alias{query_naif_switch1}
\title{Code SQL}
\usage{
query_naif_switch1(
  debut,
  fin,
  type_rx = "DENOM",
  codes,
  rx_retrospect_a_exclure = NULL,
  njours_sans_conso = 365,
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  code_list = NULL,
  code_list_filtre = "Inclusion",
  ...
)
}
\arguments{
\item{debut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ} (une seule valeur).}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ} (une seule valeur).}

\item{codes}{Traitement(s) d'intérêt(s). Vecteur comprenant le ou les codes d'analyse au format numérique, sans zéros.}

\item{rx_retrospect_a_exclure}{Traitement(s) à inclure dans la période rétrospective. Voir \emph{Details}. Un individu qui a au moins un traitement durant la période rétrospective ne sera pas considéré comme \emph{naïf} ou \emph{switch}.}

\item{njours_sans_conso}{Nombre de jours qu'un individu ne doit pas avoir reçu de traitements avant sa date de référence (index) pour être considéré \emph{naïf} ou \emph{switch}.}

\item{code_serv}{Vecteur de type \code{character} comprenant le ou les codes de service (\code{SMED_COD_SERV_1}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de service \code{code_serv}. Inscrire \code{code_serv = NULL} s'il n'y a pas de filtre à appliquer.}

\item{code_list}{Vecteur de type \code{character} comprenant le ou les codes de catégories de liste de médicaments (\code{SMED_COD_CATG_LISTE_MED}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de catégories de liste de médicaments \code{code_list}. Inscrire \code{code_list = NULL} s'il n'y a pas de filtre à appliquer.}

\item{type_Rx}{Indique le type de code analysé :
\itemize{
\item \code{'DENOM'} : Code de dénomination commune (\code{SMED_COD_DENOM_COMNE}).
\item \code{'DIN'} : Code d'identification du médicament (\code{SMED_COD_DIN}).
}}
}
\value{
Chaîne de caractères à utiliser dans une requête SQL.
}
\description{
Générateur de code SQL pour la méthode \code{naif_switch1}.
}
\details{
\strong{\code{rx_retrospect_a_exclure} :}\cr
La période rétrospective est construite à partir des dates de références (index) et de l'argument \code{njours_sans_conso} : \code{[INDEX - njours_sans_conso; INDEX - 1]}.\cr\cr
\strong{\code{code_serv_filtre}, \code{code_list_filtre} :}\cr
\code{'Exclusion'} inclus les \code{NULL}\cr
\code{'Inclusion'} exclus les \code{NULL}.
}
\examples{
### DENOM
cat(query_naif_switch1(
  debut = '2019-01-01', fin = '2019-12-31',
  type_rx = 'DENOM', codes = 47946,
  rx_retrospect_a_exclure = NULL, njours_sans_conso = 365,
  code_serv = '1', code_serv_filtre = 'Exclusion',
  code_list = NULL, code_list_filtre = 'Inclusion'
))

### DIN
cat(query_naif_switch1(
  debut = '2019-01-01', fin = '2019-12-31',
  type_rx = 'DIN', codes = 2257238,
  rx_retrospect_a_exclure = NULL, njours_sans_conso = 365,
  code_serv = '1', code_serv_filtre = 'Exclusion',
  code_list = NULL, code_list_filtre = 'Inclusion'
))

### Plusieurs codes
cat(query_naif_switch1(
  debut = '2019-01-01', fin = '2019-12-31',
  type_rx = 'DENOM', codes = c(39, 47135, 48222),
  rx_retrospect_a_exclure = NULL, njours_sans_conso = 365,
  code_serv = '1', code_serv_filtre = 'Exclusion',
  code_list = NULL, code_list_filtre = 'Inclusion'
))

### rx_retrospect_a_exclure
# Plusieurs, dont celui inclus dans 'codes'
cat(query_naif_switch1(
  debut = '2019-01-01', fin = '2019-12-31',
  type_rx = 'DENOM', codes = 47946,
  rx_retrospect_a_exclure = c(47946, 48222, 48259), njours_sans_conso = 365,
  code_serv = '1', code_serv_filtre = 'Exclusion',
  code_list = NULL, code_list_filtre = 'Inclusion'
))
# Plusieurs, sans celui inclus dans 'codes'
cat(query_naif_switch1(
  debut = '2019-01-01', fin = '2019-12-31',
  type_rx = 'DENOM', codes = 47946,
  rx_retrospect_a_exclure = c(48222, 48259, 43012), njours_sans_conso = 365,
  code_serv = '1', code_serv_filtre = 'Exclusion',
  code_list = NULL, code_list_filtre = 'Inclusion'
))

### Exclusion vs Inclusion
cat(query_naif_switch1(
  debut = '2019-01-01', fin = '2019-12-31',
  type_rx = 'DENOM', codes = 47946,
  rx_retrospect_a_exclure = NULL, njours_sans_conso = 365,
  code_serv = c('1', 'AD'), code_serv_filtre = 'Exclusion',
  code_list = c('40', '41'), code_list_filtre = 'Inclusion'
))
}
