% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch1.var_type_rx}
\alias{query_naif_switch1.var_type_rx}
\title{query_naif_switch1}
\usage{
query_naif_switch1.var_type_rx(type_rx)
}
\description{
Déterminer le nom de la variable de la vue \emph{V_DEM_PAIMT_MED_CM} selon \code{type_rx}.
}
\keyword{internal}
