% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch1.verif_args}
\alias{query_naif_switch1.verif_args}
\title{query_naif_switch1}
\usage{
query_naif_switch1.verif_args(
  debut,
  fin,
  type_Rx,
  codes,
  group_by,
  type_Rx_retro,
  code_serv,
  code_serv_filtre,
  code_list,
  code_list_filtre,
  age_date
)
}
\description{
Vérification des arguments
}
\keyword{internal}
