% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch1.where_code_list}
\alias{query_naif_switch1.where_code_list}
\title{query_naif_switch1}
\usage{
query_naif_switch1.where_code_list(code_list, code_list_filtre, cte = "")
}
\description{
Commande SQL du \code{where} pour les codes de catégorie de liste de médicament.
}
\keyword{internal}
