% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch1.where_code_serv}
\alias{query_naif_switch1.where_code_serv}
\title{query_naif_switch1}
\usage{
query_naif_switch1.where_code_serv(code_serv, code_serv_filtre, cte = "")
}
\description{
Commande SQL du \code{where} pour les codes de services.
}
\keyword{internal}
