% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.from}
\alias{query_stat_gen1.from}
\title{query_stat_gen1}
\usage{
query_stat_gen1.from(group_by)
}
\description{
Section \emph{from}.
}
\details{
Effectuer un \emph{left join} avec \emph{V_FICH_ID_BEN_CM} si on demande l'âge à une date précise.
}
\keyword{internal}
