% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.group_order_by}
\alias{query_stat_gen1.group_order_by}
\title{query_stat_gen1}
\usage{
query_stat_gen1.group_order_by(type_Rx, group_by, lvl = 0)
}
\description{
Section \emph{group by} et \emph{order by}. Générer le code SQL en fonction des variables demandées dans le \emph{group by}.
}
\keyword{internal}
