% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.sel_debut}
\alias{query_stat_gen1.sel_debut}
\title{query_stat_gen1}
\usage{
query_stat_gen1.sel_debut(debut)
}
\description{
Section \emph{select}, \emph{DATE_DEBUT}.
}
\keyword{internal}
