% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.sel_fin}
\alias{query_stat_gen1.sel_fin}
\title{query_stat_gen1}
\usage{
query_stat_gen1.sel_fin(fin)
}
\description{
Section \emph{select}, \emph{DATE_FIN}.
}
\keyword{internal}
