% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.sel_group_by}
\alias{query_stat_gen1.sel_group_by}
\title{query_stat_gen1}
\usage{
query_stat_gen1.sel_group_by(group_by, age_date, lvl = 1)
}
\description{
Section \emph{select}, colonnes à inscrire selon le \emph{group_by}.
}
\keyword{internal}
