% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.verif_args}
\alias{query_stat_gen1.verif_args}
\title{query_stat_gen1}
\usage{
query_stat_gen1.verif_args(
  debut,
  fin,
  type_Rx,
  codes,
  group_by,
  code_serv,
  code_serv_filtre,
  code_list,
  code_list_filtre,
  age_date
)
}
\description{
Vérification des arguments.
}
\keyword{internal}
