% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.where_code_list}
\alias{query_stat_gen1.where_code_list}
\title{query_stat_gen1}
\usage{
query_stat_gen1.where_code_list(code_list, code_list_filtre, lvl = 1)
}
\description{
Section \emph{where}, \emph{SMED_COD_CATG_LISTE_MED}.
}
\keyword{internal}
