% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.where_code_serv}
\alias{query_stat_gen1.where_code_serv}
\title{query_stat_gen1}
\usage{
query_stat_gen1.where_code_serv(code_serv, code_serv_filtre, lvl = 1)
}
\description{
Section \emph{where}, \emph{SMED_COD_SERV_1}.
}
\keyword{internal}
