% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.where_codes}
\alias{query_stat_gen1.where_codes}
\title{query_stat_gen1}
\usage{
query_stat_gen1.where_codes(type_Rx, codes, lvl = 1)
}
\description{
Section \emph{where}, codes Rx.
}
\keyword{internal}
