% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1.where_dates}
\alias{query_stat_gen1.where_dates}
\title{query_stat_gen1}
\usage{
query_stat_gen1.where_dates(debut, fin)
}
\description{
Section \emph{where}, \emph{SMED_DAT_SERV}. Filtrer les dates de services selon la période d'étude.
}
\keyword{internal}
