% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_DIAGN_SEJ_HOSP_CM.R
\encoding{UTF-8}
\name{query_V_DIAGN_SEJ_HOSP_CM.date_dx_var}
\alias{query_V_DIAGN_SEJ_HOSP_CM.date_dx_var}
\title{query_V_DIAGN_SEJ_HOSP_CM}
\usage{
query_V_DIAGN_SEJ_HOSP_CM.date_dx_var(date_dx_var)
}
\value{
"SHOP_DAT_ADMIS_SEJ_HOSP" ou "SHOP_DAT_DEPAR_SEJ_HOSP"
}
\description{
Retourne le nom de la variable à utiliser selon la valeur de la variable \code{date_dx_var}.
}
\keyword{internal}
