% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_DIAGN_SEJ_HOSP_CM.R
\encoding{UTF-8}
\name{query_V_DIAGN_SEJ_HOSP_CM}
\alias{query_V_DIAGN_SEJ_HOSP_CM}
\title{Code SQL}
\usage{
query_V_DIAGN_SEJ_HOSP_CM(
  debut,
  fin,
  diagn,
  date_dx_var = "admis",
  typ_diagn = c("A", "P", "S", "D")
)
}
\arguments{
\item{debut}{Date de début de la période d'étude.}

\item{fin}{Date de fin de la période d'étude.}

\item{diagn}{\code{list} indiquant les codes de diagnostics \emph{CIM9} et \emph{CIM10} à extraire.\cr \code{list(CIM9=c(...),CIM10=c(...))}.}

\item{date_dx_var}{\verb{'admis} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date de diagnostic pour l'étude.}

\item{typ_diagn}{Type de disgnostic permettant de préciser le genre de diagnostic posé pendant le séjour hospitalier. \code{A = Admission}, \code{D = Décès}, \code{P = Principal} et \code{S = Secondaire}. Voir la variable \code{SHOP_TYP_DIAGN_SEJ_HOSP} de la vue \code{V_DIAGN_SEJ_HOSP_CM}.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
Générateur de code SQL pour l'extraction de diagnostics de la vue \code{V_DIAGN_SEJ_HOSP_CM}.
}
\keyword{internal}
