% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_DIAGN_SEJ_HOSP_CM.R
\encoding{UTF-8}
\name{query_V_DIAGN_SEJ_HOSP_CM.where_diagn}
\alias{query_V_DIAGN_SEJ_HOSP_CM.where_diagn}
\title{query_V_DIAGN_SEJ_HOSP_CM}
\usage{
query_V_DIAGN_SEJ_HOSP_CM.where_diagn(diagn)
}
\description{
Section \code{where} du code SQL où on demande les codes de diagnostics.
}
\keyword{internal}
