% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_EPISO_SOIN_DURG_CM.R
\encoding{UTF-8}
\name{query_V_EPISO_SOIN_DURG_CM}
\alias{query_V_EPISO_SOIN_DURG_CM}
\title{Code SQL}
\usage{
query_V_EPISO_SOIN_DURG_CM(debut, fin, diagn, date_dx_var)
}
\arguments{
\item{debut}{Date de début de la période d'étude.}

\item{fin}{Date de fin de la période d'étude.}

\item{diagn}{\code{vector} indiquant les codes de diagnostics \emph{CIM9} et/ou \emph{CIM10} à extraire.}

\item{date_dx_var}{\verb{'admis} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date de diagnostic pour l'étude.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
Générateur de code SQL pour l'extraction de diagnostics de la vue \code{V_EPISO_SOIN_DURG_CM}.
}
\keyword{internal}
