% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_NA_in_dt.R
\encoding{UTF-8}
\name{replace_NA_in_dt}
\alias{replace_NA_in_dt}
\title{Utils}
\usage{
replace_NA_in_dt(dt, by)
}
\arguments{
\item{dt}{Tableau contenant des \code{NA}s.}

\item{by}{Valeur de remplacement.}
}
\description{
Remplace les \code{NA}s dans un tableau par \code{by}.
}
\keyword{internal}
