% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlibs.R
\encoding{UTF-8}
\name{rlibs}
\alias{rlibs}
\title{Utils}
\usage{
rlibs(x = "PROD")
}
\arguments{
\item{x}{\code{"PROD"} ou \code{"DEV"}. Par défaut \code{"PROD"}.}
}
\value{
\code{.libPaths()}
}
\description{
Modifier le répertoire des packages utilisés entre les environnements de PROD ou DEV.
}
\examples{
\dontrun{
rlibs("PROD")
rlibs("DEV")
}
}
