% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLS_convert.R
\name{RLS_convert}
\alias{RLS_convert}
\title{Convertion RLS}
\usage{
RLS_convert(dt, rls_colname)
}
\arguments{
\item{dt}{Table pouvant contenir des RLS à convertir.}

\item{rls_colname}{Nom de la colonne contenant les codes de RLS.}
}
\value{
\code{data.table}
}
\description{
Le projet de loi n\ifelse{html}{\out{<sup>o</sup>}}{\eqn{^{o}}}10 a modifié la plupart des codes RLS (voir la table \code{RLS_tab_convert}). Cette fonction permet de convertir les RLS d'une table si elle contient des codes qui existaient avant la loi 10.
}
\examples{
dt = data.frame(id = 1:5,
                age = c(45, 65, 78, 15, 35),
                sexe = c("F", "M", "F", "F", "M"),
                rls = c(1204, 215, 1611, 1503, 1610))
rls_colname = "rls"
dt_convert <- RLS_convert(dt, rls_colname = "rls")
}
