% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNA.R
\encoding{UTF-8}
\name{rmNA}
\alias{rmNA}
\title{Utils}
\usage{
rmNA(x)
}
\arguments{
\item{x}{Vecteur.}
}
\description{
Supprime les \code{NA}s du vecteur. Renvoie \code{NULL} si aucune valeur.
}
\keyword{internal}
