% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_adr.R
\encoding{UTF-8}
\name{SQL_adr}
\alias{SQL_adr}
\title{SQL}
\usage{
SQL_adr(
  user,
  t1 = NULL,
  t2 = NULL,
  cols = c("BENF_NO_INDIV_BEN_BANLS", "LGEO_COD_TERRI_CLSC", "LGEO_COD_TERRI_RLS",
    "LGEO_COD_RSS", "BENF_DD_ADR_BEN", "BENF_DF_ADR_BEN"),
  cols_name = list(BENF_NO_INDIV_BEN_BANLS = "ID", LGEO_COD_TERRI_CLSC = "CLSC",
    LGEO_COD_TERRI_RLS = "RLS", LGEO_COD_RSS = "RSS", BENF_DD_ADR_BEN = "DateDebut",
    BENF_DF_ADR_BEN = "DateFin"),
  cols_type = list(BENF_NO_INDIV_BEN_BANLS = "int", LGEO_COD_TERRI_CLSC = "int",
    LGEO_COD_TERRI_RLS = "int", LGEO_COD_RSS = "int", BENF_DD_ADR_BEN = "date",
    BENF_DF_ADR_BEN = "date"),
  select_ids = NULL,
  save_file = NULL,
  save_ext = "rds",
  ...
)
}
\arguments{
\item{user}{Nom d'usager SQL.}

\item{t1, t2}{Dates au format "AAAA-MM-JJ".\cr\code{t1=NULL} & \code{t2=NULL} : Toutes les observations.\cr\code{t1='AAAA-MM-JJ'} & \code{t2=NULL} : Observations où la date de début de la période d'application est plus petite ou égale à \code{t1} (\code{BENF_DD_ADR_BEN <= t1}).\cr\verb{t1='AAAA-MM-JJ} & \verb{t2='AAAA-MM-JJ} : Observations qui chevauchent l'intervalle \verb{[t1, t2]} (\code{BENF_DD_ADR_BEN<=t2} et \code{BENF_DF_ADR_BEN>=t1}).}

\item{cols}{Vecteur des colonnes à extraire. \code{NULL} implique toutes les colonnes. Voir \emph{Details}.}

\item{cols_name}{\code{list(nom_initial = nouveau_nom)}. Permet de modifier le nom des variables. \code{NULL} implique que l'on conserve les noms initiaux.}

\item{cols_type}{\code{list(nom_initial = type)}. Permet de modifier/forcer la classe des variables. \code{"int"=integer}, \code{"num"=numeric}, \code{"chr"=character}, \code{"date"=Date} et \code{"logical"=logical}. \code{NULL} implique aucune modification, conservation des classes initiales.}

\item{select_ids}{Vecteur contenant les bénéficiaires/usagers/individus à conserver.}

\item{save_file}{\code{"repertoire/nom_fichier"} permettant d'enregistrer le dataset sur le disque dur. Exemple : \code{"dir1/dir2/dir3/nom_du_fichier"}.}

\item{save_ext}{Extension du fichier à enregistrer. \code{"rds"}, \code{"csv"} ou \code{"txt"}.}
}
\description{
Vue : I_BENF_ADR_CM\cr
Différentes informations se rapportant aux adresses d'une personne assurée (bénéficiaire) à travers le temps.
}
\details{
\strong{Variables disponibles :}\cr
BENF_COD_TYP_ADR : CODE TYPE ADRESSE. Ce code indique à quelle fin sert l'adresse.\cr
BENF_DD_ADR_BEN : Date de début de la période d'application.\cr
BENF_DF_ADR_BEN : Date de fin de la période d'application.\cr
BENF_DM_ADR_BEN : DATE MISE À JOUR ADRESSE BÉNÉFICIAIRE. Date de mise à jour de l'occurrence.\cr
BENF_IND_ADR_HQ : INDICATEUR ADRESSE HORS-QUÉBEC. Ce champ indique si l'adresse inscrite est une adresse qui est hors du Québec (autre province ou hors du Canada).\cr
BENF_NO_INDIV_BEN_BANLS : Numéro identifiant de manière unique une personne physique qui peut être un bénéficiaire d'un service. Ce numéro est banalisé spécifiquement pour le Conseil du médicament.\cr
BENF_NO_INDIV_BEN_PORTR_BANLS : Numéro identifiant de manière unique une personne physique qui peut être un bénéficiaire et porteur de l'adresse d'un autre bénéficiaire. Ce numéro est banalisé spécifiquement pour le Conseil du médicament.\cr
LGEO_COD_RSS : Code de région socio-sanitaire.\cr
LGEO_COD_POSTL_PART : Code postal partiel, c'est-à-dire les 3 premiers caractères à partir de la gauche du code postal.\cr
LGEO_COD_TERRI_RLS : Code de territoire de RLS.\cr
LGEO_COD_TERRI_CLSC : Code de territoire de CLSC\cr
LGEO_COD_TERRI_RUIS : Code de territoire de RUIS\cr
LGEO_COD_LOCAL : CODE LOCALITE.\cr
LGEO_COD_MUNIC : Code de municipalité.\cr
LGEO_COD_MRC : CODE MUNICIPALITES REGIONALES DE COMTE.\cr
LGEO_COD_AIRE_DIFSN_STCAN : Code identifiant une aire de diffusion selon Statistique Canada.
}
\keyword{internal}
