% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_comorbidity.R
\name{SQL_comorbidity}
\alias{SQL_comorbidity}
\title{Comorbidity}
\usage{
SQL_comorbidity(
  conn,
  uid,
  pwd,
  dt,
  ID,
  DATE_INDEX,
  method = c("Charlson", "Elixhauser"),
  CIM = c("CIM9", "CIM10"),
  scores = "CIM10",
  lookup = 2,
  n1 = 30,
  n2 = 730,
  dt_source = c("V_DIAGN_SEJ_HOSP_CM", "V_SEJ_SERV_HOSP_CM", "V_EPISO_SOIN_DURG_CM",
    "I_SMOD_SERV_MD_CM"),
  dt_desc = list(V_DIAGN_SEJ_HOSP_CM = "MED-ECHO", V_SEJ_SERV_HOSP_CM = "MED-ECHO",
    V_EPISO_SOIN_DURG_CM = "BDCU", I_SMOD_SERV_MD_CM = "SMOD"),
  confirm_sourc = list(`MED-ECHO` = 1, BDCU = 2, SMOD = 2),
  exclu = list(obstetric_diab = TRUE, obstetric_hyper = TRUE),
  verbose = TRUE
)
}
\description{
Comorbidity
}
