% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_comorbidity_diagn.R
\name{SQL_comorbidity_diagn.I_SMOD_SERV_MD_CM}
\alias{SQL_comorbidity_diagn.I_SMOD_SERV_MD_CM}
\title{SQL_comorbidity_diagn}
\usage{
SQL_comorbidity_diagn.I_SMOD_SERV_MD_CM(
  conn,
  ids,
  diagn,
  debut,
  fin,
  diag_desc,
  sourc_desc,
  date_dx_var
)
}
\description{
Requête SQL pour extraire les diagnostics de la vue I_SMOD_SERV_MD_CM
}
\details{
L'argument \code{date_dx_var} n'est pas utilisé dans cette fonction. Il est là pour permettre une boucle avec les autres fonctions de requêtes.
}
\keyword{internal}
