% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_comorbidity_diagn.R
\encoding{UTF-8}
\name{SQL_comorbidity_diagn.select_Dx_table}
\alias{SQL_comorbidity_diagn.select_Dx_table}
\title{SQL_comorbidity_diagn}
\usage{
SQL_comorbidity_diagn.select_Dx_table(Dx_table)
}
\description{
Sélection du data contenant les codes à analyser.
}
\keyword{internal}
