% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_comorbidity_diagn.R
\name{SQL_comorbidity_diagn.V_SEJ_SERV_HOSP_CM}
\alias{SQL_comorbidity_diagn.V_SEJ_SERV_HOSP_CM}
\title{SQL_comorbidity_diagn}
\usage{
SQL_comorbidity_diagn.V_SEJ_SERV_HOSP_CM(
  conn,
  ids,
  diagn,
  debut,
  fin,
  diag_desc,
  sourc_desc,
  date_dx_var
)
}
\description{
Requête SQL pour extraire les diagnostics de la vue V_SEJ_SERV_HOSP_CM
}
\keyword{internal}
