% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{sql_connexion}
\alias{sql_connexion}
\title{Utils}
\usage{
sql_connexion(dsn, uid, pwd, encoding = "latin1")
}
\arguments{
\item{dsn}{\strong{D}ata \strong{S}ource \strong{N}ame.}

\item{uid}{User Identifier.}

\item{pwd}{Password to use.}
}
\value{
Connexion Teradata, sinon \code{NULL}.
}
\description{
Connexion entre R et SQL Teradata.
}
