% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_diagn.R
\encoding{UTF-8}
\name{SQL_diagn}
\alias{SQL_diagn}
\title{Diagnostiques}
\usage{
SQL_diagn(
  conn,
  uid,
  pwd,
  cohort,
  debut,
  fin,
  diagn_codes = Comorbidity_SQL_regex,
  dt_source = c("V_DIAGN_SEJ_HOSP_CM", "V_SEJ_SERV_HOSP_CM", "V_EPISO_SOIN_DURG_CM",
    "I_SMOD_SERV_MD_CM"),
  dt_desc = list(V_DIAGN_SEJ_HOSP_CM = "MED-ECHO", V_SEJ_SERV_HOSP_CM = "MED-ECHO",
    V_EPISO_SOIN_DURG_CM = "BDCU", I_SMOD_SERV_MD_CM = "SMOD")
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}. Évite d'utiliser les arguments \code{uid} et \code{pwd}.}

\item{uid}{Nom de l'identifiant pour la connexion SQL Teradata.}

\item{pwd}{Mot de passe associé à l'identifiant. Si \code{NULL}, le programme demande le mot passe. Cela permet de ne pas afficher le mot de passe dans un script.}

\item{cohort}{Cohorte d'étude. Vecteur comprenant les numéros d'identification des individus à conserver.}

\item{debut}{Date de début de la période d'étude au format \code{'AAAA-MM-JJ'}.}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{diagn_codes}{\code{list} contenant les codes \emph{regex SQL} des codes de diagnostiques. Voir \emph{Details}.}

\item{dt_source}{Vecteur comprenant la ou les bases de données où aller chercher l'information. Voir \emph{Details}.}

\item{dt_desc}{\code{list} décrivant les bases de données demandées dans \code{dt_source} au format \code{list(BD = 'MaDescription')}. Voir \emph{Details}.}
}
\value{
\code{data.table} de 4 variables :
\itemize{
\item \strong{\code{ID}} : Numéro d'identification de l'usager.
\item \strong{\code{DATE_DX}} : Date de diagnostique.
\item \strong{\code{DIAGN}} : Code descriptif des diagnostiques provenant de \code{diagn_codes}.
\item \strong{\code{SOURCE}} : Indique d'où provient l'information. Une valeur parmi \code{dt_source}.
}
}
\description{
Extraction SQL des diagnostiques.
}
\details{
Pour se connecter à Teradata, utiliser \code{conn} ou la combinaison \code{uid} et \code{pwd}.
}
