% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_naif_switch1.R
\encoding{UTF-8}
\name{SQL_naif_switch1.verif_args}
\alias{SQL_naif_switch1.verif_args}
\title{SQL_naif_switch1}
\usage{
SQL_naif_switch1.verif_args(
  debut,
  fin,
  type_Rx,
  codes,
  grouper_par,
  rx_retrospect_a_exclure,
  njours_sans_conso,
  code_serv,
  code_serv_filtre,
  code_list,
  code_list_filtre
)
}
\description{
Vérification des arguments. Les arguments manquants sont vérifiés dans la fonction \code{\link{query_naif_switch1}}.
}
\keyword{internal}
