% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_obstetric.R
\name{SQL_obstetric}
\alias{SQL_obstetric}
\title{Diagn Obstetric}
\usage{
SQL_obstetric(
  conn,
  uid,
  pwd,
  cohort,
  debut,
  fin,
  CIM = c("CIM9", "CIM10"),
  dt_source = c("V_DIAGN_SEJ_HOSP_CM", "V_SEJ_SERV_HOSP_CM", "V_EPISO_SOIN_DURG_CM",
    "I_SMOD_SERV_MD_CM"),
  verbose = TRUE
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}. Évite d'utiliser les arguments \code{uid} et \code{pwd}.}

\item{uid}{Nom de l'identifiant pour la connexion SQL Teradata.}

\item{pwd}{Mot de passe associé à l'identifiant. Si \code{NULL}, le programme demande le mot passe. Cela permet de ne pas afficher le mot de passe dans un script.}

\item{cohort}{Cohorte d'étude. Vecteur comprenant les numéros d'identification des individus à conserver.}

\item{debut}{Date de début de la période d'étude au format \code{'AAAA-MM-JJ'}.}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{CIM}{\code{'CIM9'}, \code{'CIM10'} ou les deux. Permet de filtrer les codes de diagnostiques selon le numéro de révision de la \emph{Classification statistique internationale des maladies et des problèmes de santé connexes} (CIM).}

\item{dt_source}{Vecteur comprenant la ou les bases de données où aller chercher l'information. Voir \emph{Details}.}

\item{verbose}{\code{TRUE} ou \code{FALSE}. Affiche le temps qui a été nécessaire pour extraire les diagnostiques d'une source (\code{dt_source}). Utile pour suivre le déroulement de l'extraction.}
}
\value{

}
\description{
Extraction des événements obstétriques
}
