% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_stat_gen1.R
\encoding{UTF-8}
\name{SQL_stat_gen1.ajout_nom_codes}
\alias{SQL_stat_gen1.ajout_nom_codes}
\title{SQL_stat_gen1}
\usage{
SQL_stat_gen1.ajout_nom_codes(DT, group_by)
}
\description{
Ajouter le nom des médicaments selon le group_by. Voir \emph{Details}.
}
\details{
DENOM fait afficher le nom du DENOM; DIN fait afficher le nom de la marque commerciale; AHFS fait afficher le nom de la classe AHFS.
}
\keyword{internal}
