% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_stat_gen1.R
\encoding{UTF-8}
\name{SQL_stat_gen1.cols_order}
\alias{SQL_stat_gen1.cols_order}
\title{SQL_stat_gen1}
\usage{
SQL_stat_gen1.cols_order(DT, group_by)
}
\description{
Ordre des colonnes.
}
\keyword{internal}
