% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLQ_stat_gen1.R
\name{sql_stat_gen1}
\alias{sql_stat_gen1}
\title{Statistiques générales}
\usage{
sql_stat_gen1(
  conn = NULL,
  user = NULL,
  pwd = NULL,
  debut,
  fin,
  type_Rx = "DENOM",
  codes,
  groupby = NULL,
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  code_list = NULL,
  code_list_filtre = "Inclusion"
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{sql_connexion}}.}

\item{user}{Nom de l'identifiant pour la connexion SQL Teradata.}

\item{pwd}{Mot de passe associé à l'identifiant. Si \code{NULL}, le programme demande le mot passe. Cela permet de ne pas afficher le mot de passe dans un script.}

\item{debut, fin}{Date de début et de fin de la période d'étude au format \code{"AAAA-MM_JJ"}.}

\item{type_Rx}{\code{"DENOM"} ou \code{"DIN"}. Indique le type de code analysé.}

\item{codes}{Vecteur comprenant le ou les codes d'analyse au format numérique, sans zéros.}

\item{groupby}{Regrouper les résultats par :
\itemize{
\item \strong{\code{"Période"}} : périodes d'étude.
}}

\item{code_serv}{Vecteur de type \code{character} comprenant le ou les codes de service à inclure ou exclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{"Inclusion"} ou \code{"Exclusion"} des codes de service \code{code_serv}, sinon inscrire \code{NULL}.}

\item{code_list}{Vecteur de type \code{character} comprenant le ou les codes de catégories de liste de médicaments à inclure ou exclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{"Inclusion"} ou \code{"Exclusion"} des codes de catégories de liste de médicaments \code{code_list}, sinon inscrire \code{NULL}.}
}
\value{
\code{data.table}
}
\description{
Tableau indiquant les statistiques générales d'un ou de plusieurs codes de médicaments selon certains critères.
}
\details{
Utiliser soit le paramètres \code{conn} ou la combinaison \code{user} et \code{pwd}.
}
