% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_stat_gen1.R
\encoding{UTF-8}
\name{SQL_stat_gen1.select_type_rx_var}
\alias{SQL_stat_gen1.select_type_rx_var}
\title{SQL_stat_gen1}
\usage{
SQL_stat_gen1.select_type_rx_var(type_Rx)
}
\description{
Nom de la variable selon le type Rx.
}
\keyword{internal}
