% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_stats_SMED_NBR_JR_DUREE_TRAIT.R
\encoding{UTF-8}
\name{SQL_stats_SMED_NBR_JR_DUREE_TRAIT.query}
\alias{SQL_stats_SMED_NBR_JR_DUREE_TRAIT.query}
\title{SQL_stats_SMED_NBR_JR_DUREE_TRAIT}
\usage{
SQL_stats_SMED_NBR_JR_DUREE_TRAIT.query(
  yr,
  mth,
  deb_day,
  fin_day,
  by_code_serv = TRUE,
  include_dureeTx_0 = FALSE
)
}
\arguments{
\item{yr}{Année.}

\item{mth}{Mois}

\item{deb_day}{Première journée.}

\item{fin_day}{Dernière journée.}

\item{by_code_serv}{\code{TRUE} ou \code{FALSE}. Si on veut les codes de services (pour résultats par code de service).}

\item{include_dureeTx_0}{\code{TRUE} ou \code{FALSE}. Inclure les durées de traitement égale à zéro (0).}
}
\description{
Code SQL à exécuter pour chaque mois à l'intérieur de la période d'étude.
}
\keyword{internal}
