% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_usage_statistics1.R
\name{SQL_usage_statistics1}
\alias{SQL_usage_statistics1}
\title{Usage statistics}
\usage{
SQL_usage_statistics1(user, years, code_var, codes, ...)
}
\arguments{
\item{user}{Nom d'usager SQL.}

\item{years}{Année(s) à analyser. Deux valeurs à inscrire, soit la date de début et la date de fin. Par exemple de 2015 à 2020, inscrire \code{years = c(2015, 2020)}. Si seulement une année à analyser, inscrire une seule valeur : \code{years = 2015}.}

\item{code_var}{Une des trois valeurs suivante : \code{"AHFS"}, \code{"DC"} ou \code{"DIN"}.
\itemize{
\item \code{AHFS} : Code de classe AHFS.
\item \code{DC} : Code dénomination commune.
\item \code{DIN} : Code d'identification du médicament.
}}

\item{codes}{Codes à analyser.}
}
\description{
Tableau de fréquence de différentes variables provenant de la vue \emph{V_DEM_PAIMT_MED_CM}.
}
\details{
\strong{À Faire} AHFS : Attendre de voir comment on gère les classes AHFS. Selon les cas, à tester avec Teradata.
}
\author{
Guillaume Boucher
}
