% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLQ_stat_gen1.R
\name{stat_gen1_txt_query_1period}
\alias{stat_gen1_txt_query_1period}
\title{Statistiques générales}
\usage{
stat_gen1_txt_query_1period(
  debut,
  fin,
  type_Rx,
  codes,
  groupby,
  code_serv,
  code_serv_filtre,
  code_list,
  code_list_filtre
)
}
\arguments{
\item{debut, fin}{Date de début et de fin de la période d'étude au format \code{"AAAA-MM_JJ"}.}

\item{type_Rx}{\code{"DENOM"} ou \code{"DIN"}. Indique le type de code analysé.}

\item{codes}{Vecteur comprenant le ou les codes d'analyse.}

\item{groupby}{Regrouper les résultats par :
\itemize{
\item \strong{\code{"Période"}} : périodes d'étude.
}}

\item{code_serv}{Vecteur comprenant le ou les codes de service à inclure ou exclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{"Inclusion"} ou \code{"Exclusion"} des codes de service \code{code_serv}, sinon inscrire \code{NULL}.}

\item{code_list}{Vecteur comprenant le ou les codes de catégories de liste de médicaments à inclure ou exclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{"Inclusion"} ou \code{"Exclusion"} des codes de catégories de liste de médicaments \code{code_list}, sinon inscrire \code{NULL}.}
}
\description{
Code de la requête SQL - chaîne de caractères.
}
\keyword{internal}
