% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_generales_1.R
\encoding{UTF-8}
\name{stat_generales_1}
\alias{stat_generales_1}
\title{SQL requête}
\usage{
stat_generales_1(
  fin = Sys.Date(),
  typeRx = "DENOM",
  codesRx = NULL,
  catg_liste_med = c(3, 40, 41),
  code_serv = "1",
  code_serv_filtre = "Exclusion",
  grp_age = "Mineur-Majeur"
)
}
\arguments{
\item{fin}{Date de début de l'analyse.}

\item{typeRx}{Type de code de médicament. DENOM ou DIN.}

\item{codesRx}{Codes à analyser.}

\item{catg_liste_med}{Codes de catégorie de liste de médicament.}

\item{code_serv}{Codes de service à inclure ou exclure.}

\item{code_serv_filtre}{Inclusion ou Exclusion des codes de services.}

\item{grp_age}{}
}
\value{
\code{string}
}
\description{
Statistiques générales - Méthode 1
}
\keyword{internal}
