% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unaccent.R
\encoding{UTF-8}
\name{unaccent}
\alias{unaccent}
\title{Supprimer les accents}
\usage{
unaccent(text)
}
\arguments{
\item{text}{Chaîne de caractères à convertir, où l'on veut supprimer les accents.}
}
\value{
\code{text} sans accent.
}
\description{
Permet de convertir des caractères avec accents sans accent.
}
\examples{
unaccent("âàéèïì")
}
