% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{V_DEM_PAIMT_MED_CM.SMED_COD_DIN}
\alias{V_DEM_PAIMT_MED_CM.SMED_COD_DIN}
\title{Data - COD_DIN}
\format{
Tableau de 3 variables :
\describe{
\item{DIN}{Code d'identification du médicament (\code{SMED_COD_DIN}). \code{integer}.}
\item{DEBUT}{Première année où le code a été inscrit. \code{integer}.}
\item{FIN}{Dernière année où le code a été inscrit. \code{integer}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1823&NomVue=V\%5FDEM\%5FPAIMT\%5FMED\%5FCM+\%28DEMANDES+DE+PAIEMENT+\%2D+PROGRAMME+\%ABMEDICAMENT\%BB\%29}{V_DEM_PAIMT_MED_CM}.
}
\usage{
V_DEM_PAIMT_MED_CM.SMED_COD_DIN
}
\description{
Codes d'identification des médicaments (\code{SMED_COD_DIN}) qui sont présents dans la vue \code{V_DEM_PAIMT_MED_CM}.
}
\keyword{datasets}
