% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{V_DEM_PAIMT_MED_CM.SMED_COD_SERV}
\alias{V_DEM_PAIMT_MED_CM.SMED_COD_SERV}
\title{Data - COD_SERV}
\format{
Tableau de 5 variables :
\describe{
\item{COD_SERV}{Codes de services pouvant être isncrit dans les colonnes \code{SMED_COD_SERV_1}, \code{SMED_COD_SERV_2} et \code{SMED_COD_SERV3}. \code{character}.}
\item{SERV_1}{Première et dernière année que le code de service a été inscrit dans la colonne \code{SMED_COD_SERV_1}. Si \code{NA}, le code n'a jamais été inscrit dans cette colonne. \code{character}.}
\item{SERV_2}{Première et dernière année que le code de service a été inscrit dans la colonne \code{SMED_COD_SERV_2}. Si \code{NA}, le code n'a jamais été inscrit dans cette colonne. \code{character}.}
\item{SERV_3}{Première et dernière année que le code de service a été inscrit dans la colonne \code{SMED_COD_SERV_3}. Si \code{NA}, le code n'a jamais été inscrit dans cette colonne. \code{character}.}
\item{COD_SERV_DESC}{Description du code de service tirée de la variable \code{COD_SERV_DESC} de la vue \code{V_PARAM_SERV_MED}. \code{character}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1823&NomVue=V\%5FDEM\%5FPAIMT\%5FMED\%5FCM+\%28DEMANDES+DE+PAIEMENT+\%2D+PROGRAMME+\%ABMEDICAMENT\%BB\%29}{V_DEM_PAIMT_MED_CM}.
}
\usage{
V_DEM_PAIMT_MED_CM.SMED_COD_SERV
}
\description{
Codes de services (\code{SMED_COD_SERV_X}) qui sont présent dans la vue \code{V_DEM_PAIMT_MED_CM}.
}
\keyword{datasets}
