% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{V_DENOM_COMNE_MED}
\alias{V_DENOM_COMNE_MED}
\title{Data}
\format{
Tableau de 7 variables :
\describe{
\item{DENOM}{Code de dénomination commune (\code{NMED_COD_DENOM_COMNE}). \code{character}.}
\item{DATE_DEBUT}{Date à laquelle cette dénomination commune est apparue pour la première fois (\code{NMED_DD_DENOM_COMNE}). \code{Date}.}
\item{DATE_FIN}{Date à laquelle la dénomination commune a cessé d'être utilisée (\code{NMED_DF_DENOM_COMNE}). \code{Date}.}
\item{NOM_DENOM}{Nom de la dénomination commune du médicament (\code{NMED_NOM_DENOM_COMNE}). \code{character}.}
\item{NOM_DENOM_SYNON}{Synonyme du nom de la dénomination commune du médicament\cr(\code{NMED_NOM_DENOM_COMNE_SYNON}). \code{character}.}
\item{NOM_DENOM_ANGLAIS}{Nom anglais de la dénomination commune du médicament\cr(\code{NMED_NOM_ANGL_DENOM_COMNE}). \code{character}.}
\item{NOM_DENOM_SYNON_ANGLAIS}{Synonyme du nom anglais de la dénomination commune du médicament (\code{NMED_NOM_ANGL_DENOM_SYNON}). \code{character}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1082&NomVue=V\%5FDENOM\%5FCOMNE\%5FMED+\%28D\%E9nomination+commune+m\%E9dicament\%29}{V_DENOM_COMNE_MED}.
}
\usage{
data('V_DENOM_COMNE_MED')
}
\description{
Description des codes de dénomination commune.
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
