% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{V_DES_COD}
\alias{V_DES_COD}
\title{Data}
\format{
Tableau de 5 variables :
\describe{
\item{CODE}{Valeurs codifiées que peut prendre un élément (\code{CODE_VAL_COD}). \code{character}.}
\item{TYPE_CODE}{Nom identifiant un élément de données (\code{CODE_NOM_COD}). \code{character}.}
\item{CODE_DESC}{Description du code (\code{CODE_DES}). \code{character}.}
\item{DATE_DEBUT}{Date de début de la période d'application (\code{CODE_DD_DES_COD}). \code{Date}.}
\item{DATE_FIN}{Date de fin de la période d'application (\code{CODE_DF_DES_COD}). \code{Date}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=111&NomVue=V\%5FDES\%5FCOD+\%28DESCRIPTIONS+DES+CODES\%29}{V_DES_COD}.
}
\usage{
V_DES_COD
}
\description{
Domaine de valeurs pour les différents codes de l'environnement informationnel.
}
\keyword{datasets}
