% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{V_PRODU_MED.NMED_NOM_MARQ_COMRC}
\alias{V_PRODU_MED.NMED_NOM_MARQ_COMRC}
\title{Data - NOM_MARQ_COMRC}
\format{
Tableau de 5 variables :
\describe{
\item{DENOM}{Code de dénomination commune (\code{NMED_COD_DENOM_COMNE}). \code{character}.}
\item{DIN}{Code d'identification du médicament (\code{NMED_COD_DIN}). \code{integer}.}
\item{NOM_MARQ_COMRC}{Nom sous lequel est commercialisé un produit pharmaceutique\cr(\code{NMED_NOM_MARQ_COMRC}). \code{character}.}
\item{DATE_DEBUT}{Date de début d'une occurence de cette table (\code{NMED_DD_PRODU_MED}). \code{Date}.}
\item{DATE_FIN}{Date de fin d'une occurence de cette table (\code{NMED_DF_PRODU_MED}). \code{Date}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1555&NomVue=V\%5FPRODU\%5FMED+\%28Produit+m\%E9dicament\%29}{V_PRODU_MED}.
}
\usage{
V_PRODU_MED.NMED_NOM_MARQ_COMRC
}
\description{
Nom sous lequel est commercialisé un produit pharmaceutique.
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
