% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{V_PRODU_MED}
\alias{V_PRODU_MED}
\title{Domaine de valeur}
\format{
Tableau de 5 variables :
\describe{
\item{\href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=V\%5FPRODU\%5FMED+\%28Produit+m\%E9dicament\%29&NoSeqElmVue=27099&TypOrigElmVue=T&NoSeqElmOrig=5283}{NOM_MARQ_COMRC}}{Nom sous lequel est commercialisé un produit pharmaceutique.\cr - \code{DENOM} : Code de dénomination commune (\code{NMED_COD_DENOM_COMNE}). \code{character}.\cr - \code{DIN} : Code d'identification du médicament (\code{NMED_COD_DIN}). \code{integer}.\cr - \code{NOM_MARQ_COMRC} : Nom sous lequel est commercialisé un produit pharmaceutique\cr(\code{NMED_NOM_MARQ_COMRC}). \code{character}.\cr - \code{DATE_DEBUT} : Date d'entrée en vigueur de la mise à jour à laquelle est relié l'ajout ou la modification de cette occurrence (\code{NMED_DD_PRODU_MED}). \code{Date}.\cr - \code{DATE_FIN} : Date d'entrée en vigueur de la mise à jour \strong{moins un jour} de l'occurrence suivante (\code{NMED_DF_PRODU_MED}). \code{Date}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1555&NomVue=V\%5FPRODU\%5FMED+\%28Produit+m\%E9dicament\%29}{Dictionnaire EI}.
}
\usage{
data('V_PRODU_MED')
}
\description{
Produit qui peut faire l'objet d'une facturation. Règle générale, c'est un médicament conçu par un fabricant.
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
