% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.R
\encoding{UTF-8}
\name{comorbidity}
\alias{comorbidity}
\title{Astuce}
\usage{
comorbidity(
  dt,
  ID,
  DIAGN,
  DATE_DX,
  SOURCE,
  n1 = 30,
  n2 = 730,
  Dx_table = "Comorbidity_Dx_CCI_INSPQ18",
  scores = "CCI_INSPQ_2018_CIM10",
  confirm_sourc = list(MEDECHO = 1, BDCU = 2, SMOD = 2),
  exclu_diagn = NULL,
  keep_confirm_data = FALSE
)
}
\arguments{
\item{dt}{Dataset ayant au moins les quatre (4) colonnes \code{ID}, \code{DIAGN}, \code{DATE_DX} et \code{SOURCE}.}

\item{ID}{Nom de la colonne indiquant le numéro de l'usager, de l'individu.}

\item{DIAGN}{Nom de la colonne indiquant le code d'un diagnostic.\cr Voir \code{names(inesss::Comorbidity_diagn_codes)}.}

\item{DATE_DX}{Nom de la colonne indiquant la date du diagnostic.}

\item{SOURCE}{Nom de la colonne indiquant la provenance du diagnostic.}

\item{n1, n2}{Nombre de jours dans le but de construire l'intervalle \verb{[n1,n2]}. Pour qu'un code de diagnostic soit confirmé, il faut que \emph{DIAGN{i}} soit suivi de \emph{DIAGN{j}} (où i < j) et que le nombre de jours entre les deux soit dans l'intervalle \verb{[n1,n2]}.}

\item{Dx_table}{\code{list} personnelle contenant les codes de diagnostics ou nom du dataset (de la librairie \emph{inesss}) contenant la liste des codes de diagnostics à l'étude.
\itemize{
\item \code{'Combine_Dx_CCI_INSPQ18'}
\item \code{'Charlson_Dx_CCI_INSPQ18'}
\item \code{'Elixhauser_Dx_CCI_INSPQ18'}
\item \code{'Charlson_Dx_UManitoba16'}
}}

\item{scores}{Nom de la table à utiliser pour le calcul des indicateurs. Voir les éléments de la liste \code{ComorbidityWeights}.
\itemize{
\item \code{'CCI_INSPQ_2018_CIM9'}
\item \code{'CCI_INSPQ_2018_CIM10'}
\item \code{'UManitoba_2016'}
}}

\item{confirm_sourc}{\code{list} indiquant la \emph{confiance} des \code{SOURCE}. Si une \code{SOURCE} doit être confirmée par une autre dans l'intervalle \verb{[n1,n2]}, inscrire \code{2}, sinon \code{1}. Inscrire les sources sous le format : \code{confirm_sourc = list(source1 = 1, source2 = 2, source3 = 2, ...)}. \code{confirm_sourc} doit contenir toutes les valeurs uniques de la colonne \code{SOURCE}.}

\item{exclu_diagn}{Vecteur contenant le nom du ou des diagnostics à exclure de l'analyse. Voir la liste de \code{Dx_table} pour connaître les valeurs permises.}

\item{keep_confirm_data}{\code{TRUE} ou \code{FALSE}. Place en attribut (voir fonction \code{base::attributes}) le data \code{confirm_data} qui indique la date de repérage et la date de confirmation d'un diagnostic.}
}
\value{
\code{data.table}
}
\description{
Calcul des indicateurs de \emph{Charlson}, \emph{Elixhauser} et la combinaison des deux.
}
\details{
\strong{\code{confirm_sourc} :} Dans l'exemple \code{confirm_sourc = list(source1=1, source2=2, source3=2, ...)}, la \code{source3} pourrait confirmer la \code{source2} et vice-versa.\cr\cr
\strong{\code{Dx_table} :} Il n'est pas nécessaire d'inclure un pourcentage à la fin de chaque code. Le programme le fait automatiquement s'il n'y en a pas. Par exemple inscrire le code \code{I25} et \verb{I25\%} revient au même.
}
