% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{Comorbidity_diagn_codes}
\alias{Comorbidity_diagn_codes}
\alias{Charlson_diagn_codes}
\alias{Elixhauser_diagn_codes}
\title{Data - Codes diagnostics}
\format{
\code{list(nom_diagn = list(CIM9, CIM10))}. 32 diagnostics.

\code{list(nom_diagn = list(CIM9, CIM10))}. 16 diagnostics.

\code{list(nom_diagn = list(CIM9, CIM10))}. 29 diagnostics.
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/29578951/}{Validation of the Combined Comorbidity Index of Charlson and Elixhauser to Predict 30-Day Mortality Across ICD-9 and ICD-10}. \href{https://github.com/INESSS-QC/inesss1/blob/master/inst/articles/CCI_INSPQ.pdf}{Voir PDF.}
}
\usage{
data('Comorbidity_diagn_codes')

Charlson_diagn_codes

Elixhauser_diagn_codes
}
\description{
Codes SQL regex (se terminent par un '\%') à utiliser lors de l'extraction des codes de diagnostics pour l'étude de la comorbidité.
}
\details{
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
