% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{Comorbidity_weights}
\alias{Comorbidity_weights}
\title{Poids des codes de diagnostiques}
\format{
Tableau de 4 variables et 32 observations.
\describe{
\item{DIAGN}{Identification du diagnostique (\code{chr}).}
\item{DIAGN_CODE}{Code de diagnostique utilisé utilisé dans la programmation (\code{chr}).}
\item{CIM9}{Poids utilisé pour la \emph{9e} révision du CIM (\code{int}).}
\item{CIM10}{Poids utilisé pour la \emph{10e} révision du CIM (\code{int}).}
}
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/29578951/}{Validation of the Combined Comorbidity Index of Charlson and Elixhauser to Predict 30-Day Mortality Across ICD-9 and ICD-10}. \href{Ajouter lien}{Voir PDF.}
}
\usage{
data("Comorbidity_weights") / MaVariable <- Comorbidity_weights
}
\description{
Poids des codes de diagnostiques
}
\keyword{datasets}
