% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{create_dt_data_args_query}
\alias{create_dt_data_args_query}
\title{Utils}
\usage{
create_dt_data_args_query(dt, args_list, query)
}
\arguments{
\item{dt}{Tableau des résultats.}

\item{args_list}{\code{list} contenant les arguments.}

\item{query}{Requête SQL.}
}
\description{
Combine dans un même tableau le tableau des résultats, le tableau des arguments et un exemple de code SQL
}
\keyword{internal}
