% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaire.R
\encoding{UTF-8}
\name{formulaire}
\alias{formulaire}
\title{Formulaire Interactif}
\usage{
formulaire()
}
\description{
Permets d'exécuter des requêtes Excel à partir d'un formulaire interactif. La documentation complète du formulaire, \emph{AIDE_FORMULAIRE_DATE.pdf}, est disponible \href{https://github.com/INESSS-QC/inesss1/tree/master/Documentation}{ici}.
}
\details{
\strong{Requêtes via Excel :}\cr
Il est conseillé d'utiliser les gabarits Excel pour éviter des erreurs de structures dans les tableaux d'arguments. Les fichiers Excel sont disponibles \href{https://github.com/INESSS-QC/inesss1/tree/master/Documentation/Gabarits}{ici}.
}
\keyword{internal}
