% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{from_bd.vue}
\alias{from_bd.vue}
\title{Utils}
\usage{
from_bd.vue(bd = "PROD", vue)
}
\arguments{
\item{bd}{Nom de la base de données. Souvent \code{PROD}.}

\item{vue}{Nom de la vue.}
}
\description{
FROM clause de SQL
}
