% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{I_APME_DEM_AUTOR_CRITR_ETEN_CM}
\alias{I_APME_DEM_AUTOR_CRITR_ETEN_CM}
\title{Data - Demandes d'autorisation de Patient-Médicament d'exceptions.}
\format{
\code{list}
\describe{
\item{\href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37934&TypOrigElmVue=D&NoSeqElmOrig=37196}{DES_COURT_INDCN_RECNU}}{Valeurs uniques de la description courte complète de l'indication reconnue de PME.\cr - \code{DES_COURT_INDCN_RECNU} : \code{character}.}
\item{\href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37927&TypOrigElmVue=T&NoSeqElmOrig=9127}{NO_SEQ_INDCN_RECNU_PME}}{Indique la première et la dernière année d'utilisation.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37927&TypOrigElmVue=T&NoSeqElmOrig=9127}{\code{NO_SEQ_INDCN_RECNU}} : Numéro de séquence d'indication reconnue - PME. \code{integer}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37534&TypOrigElmVue=T&NoSeqElmOrig=9381}{\code{DD_TRAIT_DEM}} : Date de début de traitement demandée. \code{character}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37537&TypOrigElmVue=T&NoSeqElmOrig=9382}{\code{DF_TRAIT_DEM}} : Date de fin de traitement demandée. \code{character}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37944&TypOrigElmVue=T&NoSeqElmOrig=9367}{\code{DD_AUTOR}} : Date de début de l'autorisation PME. \code{character}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37946&TypOrigElmVue=T&NoSeqElmOrig=9368}{\code{DF_AUTOR}} : Date de fin de l'autorisation PME. \code{character}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37943&TypOrigElmVue=T&NoSeqElmOrig=9363}{\code{DD_APLIC_AUTOR}} : Date de début de l'applicabilité de l'autorisation de PME. \code{character}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37945&TypOrigElmVue=T&NoSeqElmOrig=9364}{\code{DF_APLIC_AUTOR}} : Date de fin de l'applicabilité de l'autorisation de PME. \code{character}.\cr - \href{http://intranet/eci/eci2/asp/ECI2P06_ElmSpec.asp?Envir=PROD&min=1&max=10&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM&NoSeqElmVue=37531&TypOrigElmVue=T&NoSeqElmOrig=1872}{\code{DAT_STA_DEM}} : Date de création ou de mise à jour du statut d'une demande d'autorisation correspondant à l'attribution du dernier statut de la demande. \code{character}.}
}
}
\source{
\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=2325&NomVue=I\%5FAPME\%5FDEM\%5FAUTOR\%5FCRITR\%5FETEN\%5FCM}{Dictionnaire EI}
}
\usage{
data('I_APME_DEM_AUTOR_CRITR_ETEN_CM')
}
\description{
Data - Demandes d'autorisation de Patient-Médicament d'exceptions.
}
\keyword{datasets}
