% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_naif_switch1.R
\encoding{UTF-8}
\name{query_naif_switch1.verif_args}
\alias{query_naif_switch1.verif_args}
\title{query_naif_switch1}
\usage{
query_naif_switch1.verif_args(
  debut,
  fin,
  type_rx,
  codes,
  rx_retrospect_a_exclure,
  njours_sans_conso,
  code_serv,
  code_serv_filtre,
  code_list,
  code_list_filtre
)
}
\description{
Vérification des arguments pour la fonction \code{query_naif_switch1}.
}
\keyword{internal}
