% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1}
\alias{query_stat_gen1}
\title{Code SQL}
\usage{
query_stat_gen1(
  debut,
  fin,
  type_Rx = "DENOM",
  codes,
  group_by = NULL,
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  code_list = NULL,
  code_list_filtre = "Inclusions"
)
}
\arguments{
\item{debut}{Date de début de la période d'étude au format \code{"AAAA-MM-JJ"} (une seule valeur).}

\item{fin}{Date de fin de la période d'étude au format \code{"AAAA-MM-JJ"} (une seule valeur).}

\item{type_Rx}{\code{"DENOM"} ou \code{"DIN"}. Indique le type de code analysé.}

\item{codes}{Vecteur comprenant le ou les codes d'analyse au format numérique, sans zéros.}

\item{group_by}{Regrouper (aggréger) les résultats par :
\itemize{
\item \strong{\code{Codes}} : Résultats par code analysé.
\item \strong{\code{Teneur}} : Résultats par teneur de médicament (\code{SMED_COD_TENR_MED}) incluant les valeurs absentes.
\item \strong{\code{Format}} : Résultats par format d'acquisition du médicament (\code{SMED_COD_FORMA_ACQ_MED}) inclouant les valeurs absentes.
}}

\item{code_serv}{Vecteur de type \code{character} comprenant le ou les codes de service (\code{SMED_COD_SERV_1}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{"Inclusion"} ou \code{"Exclusion"} des codes de service \code{code_serv}, sinon inscrire \code{NULL}.}

\item{code_list}{Vecteur de type \code{character} comprenant le ou les codes de catégories de liste de médicaments (\code{SMED_COD_CATG_LISTE_MED}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{"Inclusion"} ou \code{"Exclusion"} des codes de catégories de liste de médicaments \code{code_list}, sinon inscrire \code{NULL}.}
}
\value{
Chaîne de caractères à utiliser dans une requête SQL.
}
\description{
Générateur de code SQL pour la méthode \code{stat_gen1}.
}
\details{
\strong{Méthode \code{stat_gen1} :}\cr
Statistiques descriptives tirées de la vue \code{V_DEM_PAIMT_MED_CM}.\cr\cr
\strong{\code{group_by} :} Revient à utiliser la commande \verb{group by} dans le code SQL.
}
