% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stat_gen1.R
\encoding{UTF-8}
\name{query_stat_gen1}
\alias{query_stat_gen1}
\title{Code SQL}
\usage{
query_stat_gen1(
  debut,
  fin,
  type_Rx = "DENOM",
  codes,
  group_by = "Codes",
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  code_list = NULL,
  code_list_filtre = "Inclusion",
  ...
)
}
\arguments{
\item{debut}{Date de début de la période d'étude au format \code{'AAAA-MM-JJ'} (une seule valeur).}

\item{fin}{Date de fin de la période d'étude au format \code{'AAAA-MM-JJ'} (une seule valeur).}

\item{type_Rx}{Indique le type de code analysé :
\itemize{
\item \code{'DENOM'} : Code de dénomination commune (\code{SMED_COD_DENOM_COMNE}).
\item \code{'DIN'} : Code d'identification du médicament (\code{SMED_COD_DIN}).
}}

\item{codes}{Vecteur comprenant le ou les codes d'analyse au format numérique, sans zéros.}

\item{group_by}{Regrouper (aggréger) les résultats par :
\itemize{
\item \code{'Codes'} : Résultats par code analysé.
\item \code{'Teneur'} : Résultats par teneur de médicament (\code{SMED_COD_TENR_MED}) incluant les valeurs absentes.
\item \code{'Format'} : Résultats par format d'acquisition du médicament (\code{SMED_COD_FORMA_ACQ_MED}) inclouant les valeurs absentes.
}}

\item{code_serv}{Vecteur de type \code{character} comprenant le ou les codes de service (\code{SMED_COD_SERV_1}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de service \code{code_serv}. Inscrire \code{code_serv = NULL} s'il n'y a pas de filtre à appliquer.}

\item{code_list}{Vecteur de type \code{character} comprenant le ou les codes de catégories de liste de médicaments (\code{SMED_COD_CATG_LISTE_MED}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de catégories de liste de médicaments \code{code_list}. Inscrire \code{code_list = NULL} s'il n'y a pas de filtre à appliquer.}
}
\value{
Chaîne de caractères à utiliser dans une requête SQL.
}
\description{
Générateur de code SQL pour la méthode \code{stat_gen1}.
}
\details{
\strong{Méthode \code{stat_gen1} :}\cr
Statistiques descriptives tirées de la vue \code{V_DEM_PAIMT_MED_CM}.\cr\cr
\strong{\code{group_by} :} Revient à utiliser la commande \verb{group by} dans le code SQL.\cr\cr
\strong{\code{code_serv_filtre}, \code{code_list_filtre} :}\cr
\code{"Exclusion"} : Inclus les \code{NULL}, \code{"Inclusion"} exclus les \code{NULL}.
}
\examples{
### Avantages d'utiliser cat() si c'est pour afficher du code dans la console.
# Avec cat()
cat(query_stat_gen1(
  debut = "2020-01-01", fin = "2020-12-31",
  type_Rx = "DENOM", codes = c(47092, 47135, 48222), group_by = "Codes",
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion"
))
# Sans cat()
query_stat_gen1(
  debut = "2020-01-01", fin = "2020-12-31",
  type_Rx = "DENOM", codes = c(47092, 47135, 48222), group_by = "Codes",
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion"
)

### GROUP_BY
# group_by = NULL : resultats par periode d'etude.
cat(query_stat_gen1(
  debut = "2020-01-01", fin = "2020-12-31",
  type_Rx = "DENOM", codes = c(47092, 47135, 48222),
  group_by = NULL,
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion"
))
# group_by = "Codes" : resultats par code d'analyse.
cat(query_stat_gen1(
  debut = "2020-01-01", fin = "2020-12-31",
  type_Rx = "DENOM", codes = c(47092, 47135, 48222),
  group_by = "Codes",
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion"
))
# group_by = c("Teneur", "Format") : resultats par teneur et format.
cat(query_stat_gen1(
  debut = "2020-01-01", fin = "2020-12-31",
  type_Rx = "DENOM", codes = c(47092, 47135, 48222),
  group_by = c("Teneur", "Format"),
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion"
))

### Exclusion VS Inclusion
cat(query_stat_gen1(
  debut = "2020-01-01", fin = "2020-12-31",
  type_Rx = "DENOM", codes = c(47092, 47135, 48222),
  group_by = NULL,
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion",
  code_list = c("40", "41"), code_list_filtre = "Inclusion"
))
}
