% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{RLS_convert}
\alias{RLS_convert}
\title{Data - Correspondance RLS Loi 10}
\format{
Tableau de 2 variables et 84 observations :
\describe{
\item{RLS14}{Code de RLS \strong{avant} l'adoption de la loi 10.}
\item{RLS15}{Code de RLS \strong{après} l'adoption de la loi 10.}
}
}
\source{
\href{https://publications.msss.gouv.qc.ca/msss/fichiers/statistiques/decoupage-territorial/Doc1_Correspondance_Etablissement_Public_Loi_10.xls}{Correspondance Etablissement Public Loi 10}.\cr
\href{https://github.com/INESSS-QC/inesss1/blob/master/inst/extdata/Doc1_Correspondance_Etablissement_Public_Loi_10.xls}{Fichier Excel utilisé}.
}
\usage{
data('RLS_convert') / MaVariable <- RLS_convert.
}
\description{
Établir la correspondance des RLS avant et après l'adoption de la loi 10.
}
\details{
Certains RLS ne peuvent être convertis, car leur valeur se retrouve avant et après l'adoption de la loi 10.\cr
\code{attr(RLS_convert, "RLS_exclus")} indique les quatre (4) RLS exclus : 611, 612, 1611, 1612.\cr
\code{attr(RLS_convert, "RLS_exclus_value")} renvoie un tableau indiquant les valeurs avant et après l'adoption de la loi 10 pour ces quatre (4) RLS.\cr\cr
L'attribut \code{MaJ} indique la dernière mise à jour ou la date de création du tableau.
}
\keyword{datasets}
