% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datas.R
\docType{data}
\encoding{UTF-8}
\name{RLS_list}
\alias{RLS_list}
\title{Data - Liste des RLS}
\format{
Vecteur \code{integer} de 96 nombres.
}
\usage{
data("RLS_list") / MaVariable <- RLS_list.
}
\description{
Vecteur contenant la liste des 93 RLS plus 3 valeurs utiles lors d'analyse : 1001, 1701, 1801.
}
\keyword{datasets}
