% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_comorbidity.R
\encoding{UTF-8}
\name{SQL_comorbidity}
\alias{SQL_comorbidity}
\title{Comorbidity}
\usage{
SQL_comorbidity(
  conn,
  uid,
  pwd,
  dt,
  ID,
  DATE_INDEX,
  method = c("Charlson", "Elixhauser"),
  CIM = c("CIM9", "CIM10"),
  scores = "CIM10",
  lookup = 2,
  n1 = 30,
  n2 = 730,
  dt_source = c("V_DIAGN_SEJ_HOSP_CM", "V_SEJ_SERV_HOSP_CM", "V_EPISO_SOIN_DURG_CM",
    "I_SMOD_SERV_MD_CM"),
  dt_desc = list(V_DIAGN_SEJ_HOSP_CM = "MED-ECHO", V_SEJ_SERV_HOSP_CM = "MED-ECHO",
    V_EPISO_SOIN_DURG_CM = "BDCU", I_SMOD_SERV_MD_CM = "SMOD"),
  confirm_sourc = list(`MED-ECHO` = 1, BDCU = 2, SMOD = 2),
  obstetric_exclu = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}. Évite d'utiliser les arguments \code{uid} et \code{pwd}.}

\item{uid}{Nom de l'identifiant pour la connexion SQL Teradata.}

\item{pwd}{Mot de passe associé à l'identifiant. Si \code{NULL}, le programme demande le mot passe. Cela permet de ne pas afficher le mot de passe dans un script.}

\item{dt}{Tableau ayant au moins deux colonnes : \code{ID} et \code{DATE_INDEX}.}

\item{ID}{Nom de la colonne contenant l’identifiant unique de l’usager.}

\item{DATE_INDEX}{Nom de la colonne contenant la date index de chaque usager.}

\item{method}{Extraire les diagnostiques associés aux méthodes de calcul \emph{Charlson}, \emph{Elixhauser} ou les deux.}

\item{CIM}{\code{'CIM9'}, \code{'CIM10'} ou les deux. Permet de filtrer les codes de diagnostiques selon le numéro de révision de la \emph{Classification statistique internationale des maladies et des problèmes de santé connexes} (CIM).}

\item{scores}{\code{'CIM9'} ou \code{'CIM10'}. Nom de la colonne du dataset \code{Comorbidity_weights} à utiliser pour le calcul des indicateurs.}

\item{lookup}{Nombre entier. Années à analyser avant la date index de chaque individu.}

\item{n1}{Nombre de jours dans le but de construire l'intervalle \verb{[n1,n2]}. Pour qu'un code de diagnostique soit confirmé, il faut que \emph{DIAGN{i}} soit suivi de \emph{DIAGN{j}} (où i < j) et que le nombre de jours entre les deux soit dans l'intervalle \verb{[n1,n2]}.}

\item{n2}{Nombre de jours dans le but de construire l'intervalle \verb{[n1,n2]}. Pour qu'un code de diagnostique soit confirmé, il faut que \emph{DIAGN{i}} soit suivi de \emph{DIAGN{j}} (où i < j) et que le nombre de jours entre les deux soit dans l'intervalle \verb{[n1,n2]}.}

\item{dt_source}{Vecteur comprenant la ou les bases de données où aller chercher l'information. Voir \emph{Details}.}

\item{dt_desc}{\code{list} décrivant les bases de données demandées dans \code{dt_source} au format \code{list(BD = 'MaDescription')}. Voir \emph{Details}.}

\item{confirm_sourc}{\code{list} indiquant la \emph{confiance} des \code{SOURCE}. Si une \code{SOURCE} doit être confirmé par un autre code dans l'intervalle \verb{[n1,n2]}, inscrire \code{2}, sinon \code{1}. Inscrire les sources sous le format : \code{confirm_sourc = list(source1 = 1, source2 = 2, source3 = 2, ...)}. \code{confirm_sourc} doit contenir toutes les valeurs uniques de la colonne \code{SOURCE}.}

\item{obstetric_exclu}{\code{TRUE} ou \code{FALSE}. Si l'on doit exclure (\code{TRUE}) les diabètes et les hypertensions de type gestationnel. Voir Détails.}

\item{verbose}{\code{TRUE} ou \code{FALSE}. Affiche le temps qui a été nécessaire pour extraire les diagnostiques d'une source (\code{dt_source}). Utile pour suivre le déroulement de l'extraction.}
}
\value{
\code{data.table} :
\itemize{
\item \code{ID} : Colonne contenant l’identifiant unique de l’usager.
\item \code{Charlson} : Indicateur, seulement si \code{method} contient \code{'Charlson'}.
\item \code{Elixhauser} : Indicateur, seulement si \code{method} contient \code{'Elixhauser'}.
\item \code{Combined} : Indicateur, seulement si \code{method} contient \code{'Charlson'} et \code{'Elixhauser'}.
\item Tous les diagnostiques ainsi que leur poids (score).
}
}
\description{
Extraction des codes de diagnostiques CIM pour ensuite calculer les indicateurs de Charlson et Elixhauser.
}
\details{
\strong{\code{conn}, \code{uid}, \code{pwd} :} Pour se connecter à Teradata, utiliser \code{conn} ou la combinaison \code{uid} et \code{pwd}.\cr\cr
\strong{\code{dt} :} Si un \code{ID} a plus d'une date index, seule la première, la plus ancienne, sera conservée.\cr\cr
\strong{\code{obstetric_exclu} :} Lorsqu'un cas de diabète ou d'hypertension a lieu 120 jours avant ou 180 jours après un évènement obstétrique, on les considère de type gestationnel. Ces cas sont alors exclus de l'analyse.
}
