% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_comorbidity_diagn.R
\encoding{UTF-8}
\name{SQL_comorbidity_diagn}
\alias{SQL_comorbidity_diagn}
\title{Astuce}
\usage{
SQL_comorbidity_diagn(
  conn = SQL_connexion(),
  cohort,
  debut,
  fin,
  Dx_table = "Combine_Dx_CCI_INSPQ18",
  CIM = c("CIM9", "CIM10"),
  dt_source = c("V_DIAGN_SEJ_HOSP_CM", "V_SEJ_SERV_HOSP_CM", "V_EPISO_SOIN_DURG_CM",
    "I_SMOD_SERV_MD_CM"),
  dt_desc = list(V_DIAGN_SEJ_HOSP_CM = "MEDECHO", V_SEJ_SERV_HOSP_CM = "MEDECHO",
    V_EPISO_SOIN_DURG_CM = "BDCU", I_SMOD_SERV_MD_CM = "SMOD"),
  date_dx_var = "depar",
  typ_diagn = c("A", "P", "S"),
  exclu_diagn = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}.}

\item{cohort}{Cohorte d'étude. Vecteur comprenant les numéros d'identification des individus à conserver.}

\item{debut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{Dx_table}{Nom du dataset contenant la liste des codes de diagnostics à l'étude.
\itemize{
\item \code{'Combine_Dx_CCI_INSPQ18'}
\item \code{'Charlson_Dx_CCI_INSPQ18'}
\item \code{'Elixhauser_Dx_CCI_INSPQ18'}
\item \code{'Charlson_Dx_UManitoba16'}
}}

\item{CIM}{\code{'CIM9'}, \code{'CIM10'} ou les deux. Permet de filtrer les codes de diagnostics selon le numéro de révision de la \emph{Classification statistique internationale des maladies et des problèmes de santé connexes} (CIM).}

\item{dt_source}{Vecteur comprenant la ou les bases de données où aller chercher l'information. Voir \emph{Details}.}

\item{dt_desc}{\code{list} décrivant les bases de données demandées dans \code{dt_source} au format \code{list(BD = 'MaDescription')}. Voir \emph{Details}.}

\item{date_dx_var}{\verb{'admis} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date de diagnostic pour l'étude dans les vues V_DIAGN_SEJ_HOSP_CM, V_SEJ_SERV_HOSP_CM et V_EPISO_SOIN_DURG_CM.}

\item{typ_diagn}{Type de disgnostic permettant de préciser le genre de diagnostic posé pendant le séjour hospitalier. \code{A = Admission}, \code{D = Décès}, \code{P = Principal} et \code{S = Secondaire}. Voir la variable \code{SHOP_TYP_DIAGN_SEJ_HOSP} de la vue \code{V_DIAGN_SEJ_HOSP_CM}.}

\item{exclu_diagn}{Vecteur contenant le nom du ou des diagnostics à exclure de l'analyse. Voir la liste de \code{Dx_table} pour connaître les valeurs permises.}

\item{verbose}{\code{TRUE} ou \code{FALSE}. Affiche le temps qui a été nécessaire pour extraire les diagnostics d'une source (\code{dt_source}). Utile pour suivre le déroulement de l'extraction.}
}
\value{
\code{data.table} de 4 variables :
\itemize{
\item \strong{\code{ID}} : Numéro d'identification de l'usager.
\item \strong{\code{DATE_DX}} : Date de diagnostic.
\item \strong{\code{DIAGN}} : Code descriptif des diagnostics provenant de \code{diagn_codes}.
\item \strong{\code{SOURCE}} : Indique d'où provient l'information. Une valeur parmi \code{dt_source}.
}
}
\description{
Extraction SQL des diagnostics pour l'étude de la comorbidité.
}
\details{
\strong{\code{dt_source} :}
\itemize{
\item \href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=6721&NomVue=V\%5FDIAGN\%5FSEJ\%5FHOSP\%5FCM+\%28Diagnostic+s\%E9jour+hospitalier\%29}{\code{V_DIAGN_SEJ_HOSP_CM}} : Cette structure contient tous les diagnostics associés à un séjour hospitalier.
\item \href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=6724&NomVue=V\%5FSEJ\%5FSERV\%5FHOSP\%5FCM+\%28S\%E9jour+service+hospitalier\%29}{\code{V_SEJ_SERV_HOSP_CM}} : Cette structure contient les séjours dans un service effectués par l'individu hospitalisé.
\item \href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=6687&NomVue=V\%5FEPISO\%5FSOIN\%5FDURG\%5FCM+\%28\%C9pisodes+de+soins+en+D\%E9partement+d\%27urgence\%29}{\code{V_EPISO_SOIN_DURG_CM}} : Cette structure contient les épisodes de soins des départements d'urgence de la province.
\item \href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1797&NomVue=I\%5FSMOD\%5FSERV\%5FMD\%5FCM}{\code{I_SMOD_SERV_MD_CM}} : Cette vue retourne différentes informations se rapportant aux Services rendus à l'acte par des médecins.
}
}
