% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{sql_connexion}
\alias{sql_connexion}
\title{Utils}
\usage{
sql_connexion(uid, pwd = NULL, dsn = "PEI_PRD", encoding = "latin1")
}
\arguments{
\item{uid}{Identifiant.}

\item{pwd}{Mot de passe. Si \code{NULL}, le mot de passe est demandé lors de l'exécution.}

\item{dsn}{\strong{D}ata \strong{S}ource \strong{N}ame. Par défaut \code{"PEI_PRD"}.}

\item{encoding}{\code{"latin1"} ou \code{"UTF-8"}. Encodage de la base de données. Par défaut \code{"latin1"}.}
}
\value{
Connexion Teradata, sinon \code{NULL}.
}
\description{
Connexion entre R et SQL Teradata.
}
\details{
Voir \code{\link[odbc]{dbConnect}}.
}
