% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_naif_switch1.R
\encoding{UTF-8}
\name{SQL_naif_switch1}
\alias{SQL_naif_switch1}
\title{Astuce}
\usage{
SQL_naif_switch1(
  conn = SQL_connexion(),
  debut,
  fin,
  type_Rx = "DENOM",
  codes,
  group_by = "DENOM",
  type_Rx_retro = NULL,
  rx_retrospect_a_exclure = NULL,
  njours_sans_conso = 365,
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  code_list = NULL,
  code_list_filtre = "Inclusion",
  age_date = NULL,
  ...
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}.}

\item{debut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ} (une seule valeur).}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ} (une seule valeur).}

\item{type_Rx}{Type de code à analyser. Une valeur parmi :
\itemize{
\item \code{'DENOM'} : Code de dénomination commune (\code{SMED_COD_DENOM_COMNE}).
\item \code{'DIN'} : Code d'identification du médicament (\code{SMED_COD_DIN}).
}}

\item{codes}{Le ou les codes à analyser. Voir \emph{Details}.}

\item{group_by}{Regrouper (aggréger) les résultats par :
\itemize{
\item \code{'AHFS'} : Résultats par code de classe AHFS.
\item \code{'DENOM'} : Résultats par code de dénomination commune.
\item \code{'DIN'} : Résultats par code d'identification du médicament.
\item \code{'CodeList'} : Résultats par code de catégories de liste de médicaments.
\item \code{'CodeServ'} : Résultats par code de service.
\item \code{'Teneur'} : Résultats par teneur du médicament.
\item \code{'Format'} : Résultats par format d'acquisition du médicament.
\item \code{'Age'} : Résultats par âge à une date précise. Voir argument \code{age_date}. L'âge est calculé à partir de la date de naissance disponible dans la vue \code{V_FICH_ID_BEN_CM}.
}}

\item{type_Rx_retro}{Type de code à exclure. Si \code{NULL}, prend la valeur de \code{type_Rx}. Une valeur parmi :
\itemize{
\item \code{'AHFS'} : Code identifiant la classe de médicaments telle que déterminée par l'\emph{American Hospital Formulary Service}.
\item \code{'DENOM'} : Code de dénomination commune (\code{SMED_COD_DENOM_COMNE}).
\item \code{'DIN'} : Code d'identification du médicament (\code{SMED_COD_DIN}).
}}

\item{rx_retrospect_a_exclure}{Traitement(s) à inclure dans la période rétrospective. Voir \emph{Details}. Un individu qui a au moins un traitement durant la période rétrospective ne sera pas considéré comme \emph{naïf} ou \emph{switch}.}

\item{njours_sans_conso}{Nombre de jours qu’un individu ne doit pas avoir reçu de traitements avant sa date de référence (date index) pour être considéré \emph{naïf} ou \emph{switch}.}

\item{code_serv}{Vecteur de type \code{character} comprenant le ou les codes de service (\code{SMED_COD_SERV_1}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de service \code{code_serv}. Inscrire \code{code_serv = NULL} s'il n'y a pas de filtre à appliquer.}

\item{code_list}{Vecteur de type \code{character} comprenant le ou les codes de catégories de listes de médicaments (\code{SMED_COD_CATG_LISTE_MED}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de catégories de liste de médicaments \code{code_list}. Inscrire \code{code_list = NULL} s'il n'y a pas de filtre à appliquer.}

\item{age_date}{Date à laquelle on calcule l'âge si \code{group_by} contient \code{'Age'}. Si \code{NULL}, aura pour valeur \code{debut}.}
}
\value{
\code{data.table}
\itemize{
\item \strong{\code{DATE_DEBUT} :} Indique la ou les dates de début de la période d'étude.
\item \strong{\code{DATE_FIN} :} Indique la ou les dates de fin de la période d'étude.
\item \strong{\code{AHFS_CLA} :} Seulement si \code{group_by} contient \code{'AHFS'}. Code de la classe AHFS.
\item \strong{\code{AHFS_SCLA} :} Seulement si \code{group_by} contient \code{'AHFS'}. Code de la sous-classe AHFS.
\item \strong{\code{AHFS_SSCLA} :} Seulement si \code{group_by} contient \code{'AHFS'}. Code de la sous-sous-classe AHFS.
\item \strong{\code{NOM_AHFS} :} Seulement si \code{group_by} contient \code{'AHFS'}. Nom de la classe AHFS.
\item \strong{\code{DENOM} :} Seulement si \code{group_by} contient \code{'DENOM'}. Code de dénomination commune.
\item \strong{\code{NOM_DENOM} :} Seulement si \code{group_by} contient \code{'DENOM'}. Nom de la dénomination commune.
\item \strong{\code{DIN} :} Seulement si \code{group_by} contient \code{'DIN'}. Code d'identification du médicament.
\item \strong{\code{NOM_MARQ_COMRC} :} Seulement si \code{group_by} contient \code{'DIN'}. Nom de la marque commerciale.
\item \strong{\code{CODE_SERV} :} Seulement si \code{group_by} contient \code{'CodeServ'}. Code de service,
\item **\code{CODE_LIST} : ** Seulement si \code{group_by} contient \code{'CodeList'}. Code de catégorie de listes de médicaments.
\item \strong{\code{TENEUR} :} Seulement si \code{group_by} contient \code{'Teneur'}. Teneur du médicament.
\item \strong{\code{FORMAT_ACQ} :} Seulement si \code{group_by} contient \code{'Format'}. Format d'acquisition du médicament.
\item \strong{\code{AGE} :} Seulement si \code{group_by} contient \code{'Age'}. Age de l'individu à la date \code{age_date}.
\item \strong{\code{MNT_MED} :} Montant autorisé par la RAMQ pour le médicament ou le produit. Il comprend la part du grossiste (s'il y a lieu) et la part du manufacturier. Voir la variable \code{SMED_MNT_AUTOR_MED}.
\item \strong{\code{MNT_SERV} :} Montant de frais de service autorisé par la RAMQ à la date du service. Voir la variable \code{SMED_MNT_AUTOR_FRAIS_SERV}.
\item \strong{\code{MNT_TOT} :} Somme des variables \code{MNT_MED} et \code{MNT_SERV}.
\item \strong{\code{COHORTE} :} Nombre d'individus unique.
\item \strong{\code{NBRE_RX} :} Nombre de demandes de paiement.
\item \strong{\code{QTE_MED} :} Quantité totale des médicaments ou des fournitures dispensés. Voir la variable \code{SMED_QTE_MED}.
\item \strong{\code{DUREE_TX} :} Durée de traitement totale des prescriptions en jours. Voir la variable \code{SMED_NBR_JR_DUREE_TRAIT}.
}
}
\description{
Statistiques générales pour un ou des médicaments à partir d'une cohorte consommant ce(s) médicament(s) pour la première fois.\cr
Un individu est considéré \emph{naïf} lorsqu'il a un traitement pour la première fois et qu'il n'a jamais eu d'autres traitements \emph{de la même famille}.\cr
Un individu est considéré \emph{switch} lorsqu'il a un traitement pour la première fois, mais qu'il a eu un autre traitement dans le passé appartenant \emph{à la même famille}.\cr
Vue utilisée : \code{\href{http://intranet/eci/ECI2/ASP/ECI2P04_DescVue.asp?Envir=PROD&NoVue=1823&NomVue=V\%5FDEM\%5FPAIMT\%5FMED\%5FCM+\%28DEMANDES+DE+PAIEMENT+\%2D+PROGRAMME+\%ABMEDICAMENT\%BB\%29}{V_DEM_PAIMT_MED_CM}}.
}
\details{
\strong{rx_retrospect_a_exclure :}\cr
La période rétrospective est construite à partir des dates de références (index) et de l'argument \code{njours_sans_conso} : \code{[INDEX - njours_sans_conso; INDEX - 1]}.\cr\cr
\strong{code_serv_filtre, code_list_filtre :}\cr
\code{'Exclusion'} inclus les \code{NULL}\cr
\code{'Inclusion'} exclus les \code{NULL}.
}
