% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_reperage_cond_med.R
\encoding{UTF-8}
\name{SQL_reperage_cond_med}
\alias{SQL_reperage_cond_med}
\title{Requête complexe}
\usage{
SQL_reperage_cond_med(
  conn = SQL_connexion(),
  debut,
  fin,
  Dx_table,
  CIM = c("CIM9", "CIM10"),
  by_Dx = FALSE,
  date_dx_var = "admis",
  n1 = 30,
  n2 = 730,
  keep_all = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}.}

\item{debut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{Dx_table}{\code{list} personnelle contenant les codes de diagnostics ou nom du dataset (de la librairie \emph{inesss}) contenant la liste des codes de diagnostics à l'étude.
\itemize{
\item \code{'Combine_Dx_CCI_INSPQ18'}
\item \code{'Charlson_Dx_CCI_INSPQ18'}
\item \code{'Elixhauser_Dx_CCI_INSPQ18'}
\item \code{'Charlson_Dx_UManitoba16'}
}}

\item{CIM}{\code{'CIM9'}, \code{'CIM10'} ou les deux. Permet de filtrer les codes de diagnostics selon le numéro de révision de la \emph{Classification statistique internationale des maladies et des problèmes de santé connexes} (CIM).}

\item{by_Dx}{\code{TRUE} ou \code{FALSE}. Distinction entre les diagnostics (\code{TRUE}) ou pas (\code{FALSE}). La distinction des diagnostics implique une cohorte d'étude pour chaque élément de l'argument \code{Dx_table}, alors que \code{FALSE} tous les éléments sont réunis comme si c'était la même maladie.}

\item{date_dx_var}{\verb{'admis} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date de diagnostic pour l'étude dans les vues V_DIAGN_SEJ_HOSP_CM, V_SEJ_SERV_HOSP_CM et V_EPISO_SOIN_DURG_CM.}

\item{n1}{Nombre de jours permettant de construire l'intervalle [n1; n2] où un code de diagnostic peut en confirmer un autre.}

\item{n2}{Nombre de jours permettant de construire l'intervalle [n1; n2] où un code de diagnostic peut en confirmer un autre.}

\item{keep_all}{\code{TRUE} ou \code{FALSE}. Par défaut \code{FALSE}.\cr\code{FALSE} supprime toutes les observations où \code{DI_Finale = NA}.\cr\code{TRUE} est utile si on cherche la date la plus récente pour chaque individu.}

\item{verbose}{\code{TRUE} ou \code{FALSE}. Affiche le temps qui a été nécessaire pour extraire les diagnostics d'une source (\code{dt_source}). Utile pour suivre le déroulement de l'extraction.}
}
\value{
\code{data.table} :
\itemize{
\item \code{ID} : Identifiant de l'individu.
\item \code{DIAGN} : Nom du diagnostic. Seulement si \code{by_Dx=TRUE}.
\item \code{DI_Finale} : Date d'incidence retenue.
\item \code{DI_Hospit} : Date d'incidence d'hospitalisation.
\item \code{DI_Acte} : Date d'incidence acte.
\item \code{DC_Acte} : Date de confirmation de \code{DI_Acte}.
\item \code{D_Recent} : Date du diagnostic la plus récente \strong{sans tenir compte de l'algorithme}.
}
}
\description{
Repérage d'une condition médicale.
}
\details{
Détails à venir.
}
\examples{
### Inscription des codes de diagnostics dans Dx_table
# 1 code CIM9 ou CIM10
liste1 = list(CIM9 = "413")
liste2 = list(CIM10 = "I20")
# 1 code de chaque
liste3 = list(CIM9 = "413", CIM10 = "I20")
# Plusieurs codes qui se suivent
liste4 = list(CIM9 = 410:414, CIM10 = paste0("I", 21:25))
# Codes qui ne se suivent pas
liste5 = list(CIM9 = c(410, 412, 414), CIM10 = paste0("I", 21, 23, 25))
# Codes qui se suivent et qui ne se suivent pas
liste6 = list(CIM9 = c(405:410, 412:414),
              CIM10 = c(paste0("I", c(18:21, 23:25)),
                        paste0("K0", c(5:9)), paste0("K", 10:15)))
}
