% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_stat_gen1.R
\encoding{UTF-8}
\name{SQL_stat_gen1}
\alias{SQL_stat_gen1}
\title{Statistiques générales}
\usage{
SQL_stat_gen1(
  conn = NULL,
  uid = NULL,
  pwd = NULL,
  debut,
  fin,
  type_Rx = "DENOM",
  codes,
  group_by = "Codes",
  code_serv = c("1", "AD"),
  code_serv_filtre = "Exclusion",
  code_list = NULL,
  code_list_filtre = "Inclusion",
  ...
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}.}

\item{uid}{Nom de l'identifiant pour la connexion SQL Teradata.}

\item{pwd}{Mot de passe associé à l'identifiant. Si \code{NULL}, le programme demande le mot passe. Cela permet de ne pas afficher le mot de passe dans un script.}

\item{debut}{Date de début de la période d'étude au format \code{'AAAA-MM-JJ'} (une seule valeur).}

\item{fin}{Date de fin de la période d'étude au format \code{'AAAA-MM-JJ'} (une seule valeur).}

\item{type_Rx}{Indique le type de code analysé :
\itemize{
\item \code{'DENOM'} : Code de dénomination commune (\code{SMED_COD_DENOM_COMNE}).
\item \code{'DIN'} : Code d'identification du médicament (\code{SMED_COD_DIN}).
}}

\item{codes}{Vecteur comprenant le ou les codes d'analyse au format numérique, sans zéros.}

\item{group_by}{Regrouper (aggréger) les résultats par :
\itemize{
\item \code{'Codes'} : Résultats par code analysé.
\item \code{'Teneur'} : Résultats par teneur de médicament (\code{SMED_COD_TENR_MED}) incluant les valeurs absentes.
\item \code{'Format'} : Résultats par format d'acquisition du médicament (\code{SMED_COD_FORMA_ACQ_MED}) inclouant les valeurs absentes.
}}

\item{code_serv}{Vecteur de type \code{character} comprenant le ou les codes de service (\code{SMED_COD_SERV_1}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_serv_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de service \code{code_serv}. Inscrire \code{code_serv = NULL} s'il n'y a pas de filtre à appliquer.}

\item{code_list}{Vecteur de type \code{character} comprenant le ou les codes de catégories de liste de médicaments (\code{SMED_COD_CATG_LISTE_MED}) à exclure ou à inclure, sinon inscrire \code{NULL}.}

\item{code_list_filtre}{\code{'Inclusion'} ou \code{'Exclusion'} des codes de catégories de liste de médicaments \code{code_list}. Inscrire \code{code_list = NULL} s'il n'y a pas de filtre à appliquer.}
}
\description{
Tableau indiquant les statistiques générales d'un ou de plusieurs codes de médicaments selon certains critères.
}
\examples{
\dontrun{
### GROUP_BY
# Aucun group by
dt1 <- SQL_stat_gen1(
  conn = conn,
  debut = c("2017-01-01", "2018-01-01"), fin = c("2017-12-31", "2018-12-31"),
  type_Rx = "DENOM", codes = c(39, 47092, 48222), group_by = NULL
)
# Codes - DENOM
dt2 <- SQL_stat_gen1(
  conn = conn,
  debut = c("2017-01-01", "2018-01-01"),fin = c("2017-12-31", "2018-12-31"),
  type_Rx = "DENOM", codes = c(39, 47092, 48222), group_by = "Codes"
)
# Codes - DIN
dt3 <- SQL_stat_gen1(
  conn = conn,
  debut = c("2017-01-01", "2018-01-01"),fin = c("2017-12-31", "2018-12-31"),
  type_Rx = "DIN", codes = c(30848, 585092), group_by = "Codes"
)
# Codes et Format
dt4 <- SQL_stat_gen1(
  conn = conn,
  debut = c("2017-01-01", "2018-01-01"),fin = c("2017-12-31", "2018-12-31"),
  type_Rx = "DENOM", codes = c(39, 47092, 48222), group_by = c("Codes", "Format")
)
# Teneur
dt5 <- SQL_stat_gen1(
  conn = conn,
  debut = c("2017-01-01", "2018-01-01"),fin = c("2017-12-31", "2018-12-31"),
  type_Rx = "DENOM", codes = c(39, 47092, 48222), group_by = "Teneur"
)

### Exclusion & Inclusion
dt6 <- SQL_stat_gen1(
  conn = conn,
  debut = c("2017-01-01", "2018-01-01"),fin = c("2017-12-31", "2018-12-31"),
  type_Rx = "DENOM", codes = c(39, 47092, 48222), group_by = "Codes",
  code_serv = c("1", "AD"), code_serv_filtre = "Exclusion",
  code_list = c("03", "40", "41"), code_list_filtre = "Inclusion"
)
}
}
