% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_usage_statistics.R
\name{SQL_usage_statistics1}
\alias{SQL_usage_statistics1}
\title{Usage statistics}
\usage{
SQL_usage_statistics1(user, years, code_var, codes, ...)
}
\arguments{
\item{user}{Nom d'usager SQL.}

\item{years}{Année(s) à analyser.}

\item{code_var}{Une des trois valeurs suivante : \code{"AHFS"}, \code{"DC"} ou \code{"DIN"}.
\itemize{
\item \code{AHFS} : Code de classe AHFS.
\item \code{DC} : Code dénomination commune.
\item \code{DIN} : Code d'identification du médicament.
}}

\item{codes}{Codes à analyser.}
}
\description{
Tableau de fréquence de différentes variables provenant de la vue \emph{V_DEM_PAIMT_MED_CM}.
}
\author{
Guillaume Boucher
}
