% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLQ_stat_gen1.R
\name{stat_gen1_txt_query_1period}
\alias{stat_gen1_txt_query_1period}
\title{Statistique générale}
\usage{
stat_gen1_txt_query_1period(
  debut,
  fin,
  type_Rx,
  codes,
  code_serv,
  code_serv_filtre,
  code_list,
  code_list_filtre
)
}
\description{
Code de la requête SQL (châine de caractères)
}
\keyword{internal}
