% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLQ_stat_gen1.R
\name{stat_gen1_txt_query_1period}
\alias{stat_gen1_txt_query_1period}
\title{Statistique générale}
\usage{
stat_gen1_txt_query_1period(
  DateDebut,
  DateFin,
  TypeVar,
  Codes,
  Stats,
  GroupBy,
  ExcluCodeServ,
  CategorieListe
)
}
\arguments{
\item{DateDebut}{Date de début d'étude.}

\item{DateFin}{Date de fin d'étude}

\item{TypeVar}{'DENOM' ou 'DIN'.}

\item{Codes}{Codes à l'étude selon \code{TypeVar}.}

\item{Stats}{Vecteur indiquant les statistiques à afficher.}

\item{GroupBy}{Vecteur indiquant les \emph{group by}, par exemple Années, Code, ...}

\item{ExcluCodeServ}{Codes de service à exclure}

\item{CategorieListe}{Categorie de liste à analyser (filtre)}
}
\description{
Code de la requête SQL (châine de caractères)
}
\keyword{internal}
