% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{sunique}
\alias{sunique}
\title{Utils}
\usage{
sunique(x, decreasing = FALSE, na.last = FALSE)
}
\arguments{
\item{x}{Vecteur à trier et supprimer doublons.}

\item{decreasing}{Ordre décroissant = \code{TRUE}, sinon \code{FALSE}.}

\item{na.last}{Afficher les \code{NA} à la fin = \code{TRUE}, sinon \code{FALSE}. \code{NA} n'affiche pas les valeurs \code{NA}.}
}
\description{
Combinaison de \code{sort()} et \code{unique()}.
}
\keyword{internal}
